refactor(abc): split metrics/abc.rs into per-language submodules
The single 12.3k-line src/metrics/abc.rs was the largest file in the
repo and the dominant self-scan offender: its file-level loc.sloc of
4419 (production only — the in-file test module is exclude_tests-pruned)
sat far over the 800 limit and was baselined.

Move each `impl Abc for <Lang>Code` block and its private helpers into
src/metrics/abc/<lang>.rs, mirroring the per-language layout of
src/languages/. The four JS-family impls and the three macros they share
move together into abc/js_family.rs so every macro use stays textually
in the same file as its definition. Shared items (Stats, the Abc trait,
DeclKind, for_each_named_child) stay in the root module; DeclKind,
for_each_named_child, and the four metric-accumulator Stats fields the
per-language walkers write to are widened to pub(super) so submodules
reach them. The C-family (c/objc/mozcpp) imports cpp's three shared
helpers, now pub(super).

This is a pure code move: no metric computation changes. The full
workspace test suite passes with no snapshot drift (no submodule bump),
and the file-level loc.sloc baseline entry is retired — the nine genuine
per-function offenders (essential per-language complexity) are
path-rewritten to their new submodule files.

The 7.9k-line in-file test module stays in the root for now; partitioning
it per-language is a deferred follow-up.

Refs #969 (Finding 2)
