+++ src/metrics/abc.rs
@@ -24,18 +24,31 @@
 use std::fmt;
 
 use crate::checker::Checker;
-use crate::macros::{
-    cpp_bool_terminal_kinds, csharp_bool_terminal_kinds, csharp_paren_expr_kinds,
-    csharp_prefix_unary_expr_kinds, elixir_bool_terminal_kinds, go_bool_terminal_kinds,
-    groovy_bool_terminal_kinds, implement_metric_trait, irules_bool_terminal_kinds,
-    java_bool_terminal_kinds, javascript_bool_terminal_kinds, kotlin_bool_terminal_kinds,
-    lua_bool_terminal_kinds, mozjs_bool_terminal_kinds, perl_bool_terminal_kinds,
-    php_bool_terminal_kinds, python_bool_terminal_kinds, ruby_bool_terminal_kinds,
-    rust_bool_terminal_kinds, tcl_bool_terminal_kinds, tsx_bool_terminal_kinds,
-    typescript_bool_terminal_kinds,
-};
-use crate::node::Node;
+
+use crate::macros::implement_metric_trait;
+
 use crate::*;
+
+mod bash;
+mod c;
+mod cpp;
+mod csharp;
+mod elixir;
+mod go;
+mod groovy;
+mod irules;
+mod java;
+mod js_family;
+mod kotlin;
+mod lua;
+mod mozcpp;
+mod objc;
+mod perl;
+mod php;
+mod python;
+mod ruby;
+mod rust;
+mod tcl;
 
 /// The `ABC` metric.
 ///
@@ -103,24 +116,24 @@ use crate::*;
 /// for the Kotlin / Ruby / Elixir walker wiring.
 #[derive(Debug, Clone, PartialEq)]
 pub struct Stats {
-    assignments: f64,
+    pub(super) assignments: f64,
     assignments_sum: f64,
     assignments_min: f64,
     assignments_max: f64,
-    branches: f64,
+    pub(super) branches: f64,
     branches_sum: f64,
     branches_min: f64,
     branches_max: f64,
-    conditions: f64,
+    pub(super) conditions: f64,
     conditions_sum: f64,
     conditions_min: f64,
     conditions_max: f64,
     space_count: usize,
-    declaration: Vec<DeclKind>,
+    pub(super) declaration: Vec<DeclKind>,
 }
 
 #[derive(Debug, Clone, PartialEq)]
-enum DeclKind {
+pub(super) enum DeclKind {
     Var,
     Const,
 }
@@ -371,7 +384,7 @@ where
 // values live one level below the return statement under a list
 // wrapper. The classifier receives only named children so that
 // `,` / `;` / `(` / `)` tokens never reach it.
-fn for_each_named_child(list: &Node, conditions: &mut f64, f: fn(&Node, &mut f64)) {
+pub(super) fn for_each_named_child(list: &Node, conditions: &mut f64, f: fn(&Node, &mut f64)) {
     let mut cursor = list.cursor();
     if cursor.goto_first_child() {
         loop {
@@ -379,276 +392,6 @@ fn for_each_named_child(list: &Node, conditions: &mut f64, f: fn(&Node, &mut f64
             if child.is_named() {
                 f(&child, conditions);
             }
-            if !cursor.goto_next_sibling() {
-                break;
-            }
-        }
-    }
-}
-
-// Inspects the content of Java parenthesized expressions
-// and `Not` operators to find unary conditional expressions
-fn java_inspect_container(container_node: &Node, conditions: &mut f64) {
-    use Java::*;
-
-    let mut node = *container_node;
-    let mut node_kind = node.kind_id().into();
-
-    // Initializes the flag to true if the container is known to contain a boolean value
-    let Some(parent) = node.parent() else { return };
-    let mut has_boolean_content = match parent.kind_id().into() {
-        BinaryExpression | IfStatement | WhileStatement | DoStatement | ForStatement => true,
-        TernaryExpression => node
-            .previous_sibling()
-            .is_none_or(|prev_node| !matches!(prev_node.kind_id().into(), QMARK | COLON)),
-        _ => false,
-    };
-
-    // Looks inside parenthesized expressions and `Not` operators to find what they contain
-    loop {
-        // Checks if the node is a parenthesized expression or a `Not` operator
-        // The child node of index 0 contains the unary expression operator (we look for the `!` operator)
-        let is_parenthesised_exp = matches!(node_kind, ParenthesizedExpression);
-        let is_not_operator = matches!(node_kind, UnaryExpression)
-            && node
-                .child(0)
-                .is_some_and(|c| matches!(c.kind_id().into(), BANG));
-
-        // Stops the exploration if the node is neither
-        // a parenthesized expression nor a `Not` operator
-        if !is_parenthesised_exp && !is_not_operator {
-            break;
-        }
-
-        // Sets the flag to true if a `Not` operator is found
-        // This is used to prove if a variable or a value returned by a method is actually boolean
-        // e.g. `return (!x);`
-        if !has_boolean_content && is_not_operator {
-            has_boolean_content = true;
-        }
-
-        // Parenthesized expressions and `Not` operators nodes
-        // always store their expressions in the children nodes of index one
-        // https://github.com/tree-sitter/tree-sitter-java/blob/master/src/grammar.json#L2472
-        // https://github.com/tree-sitter/tree-sitter-java/blob/master/src/grammar.json#L2150
-        let Some(child) = node.child(1) else { break };
-        node = child;
-        node_kind = node.kind_id().into();
-
-        // Stops the exploration when the content is found. The terminal
-        // set includes `FieldAccess` (`obj.flag`), `CastExpression`
-        // (`(boolean)v`), `ArrayAccess` (`flags[0]`), and
-        // `InstanceofExpression` (`x instanceof Foo`) — every kind whose
-        // evaluated value is implicitly boolean in idiomatic Java, mirroring
-        // the C# fix in #372 (lesson #19).
-        if matches!(node_kind, java_bool_terminal_kinds!()) {
-            if has_boolean_content {
-                *conditions += 1.;
-            }
-            break;
-        }
-    }
-}
-
-fn csharp_inspect_container(container_node: &Node, conditions: &mut f64) {
-    use Csharp::*;
-
-    let mut node = *container_node;
-    let mut node_kind = node.kind_id().into();
-
-    // Seed the boolean-context flag from the parent: known-boolean
-    // contexts (loop / if / binary expression) imply the contained
-    // expression evaluates as a condition.
-    let Some(parent) = node.parent() else { return };
-    let mut has_boolean_content = match parent.kind_id().into() {
-        BinaryExpression | IfStatement | WhileStatement | DoStatement | ForStatement => true,
-        ConditionalExpression => node
-            .previous_sibling()
-            .is_none_or(|prev| !matches!(prev.kind_id().into(), QMARK | COLON)),
-        _ => false,
-    };
-
-    // Walk down through `(...)` and `!...` wrappers until we either hit
-    // the underlying operand or run out of nesting. The C# grammar
-    // aliases each of these kinds across multiple `kind_id`s
-    // (lesson #2): match every numbered variant.
-    loop {
-        let is_parens = matches!(node_kind, csharp_paren_expr_kinds!());
-        let is_not = matches!(node_kind, csharp_prefix_unary_expr_kinds!())
-            && node
-                .child(0)
-                .is_some_and(|c| matches!(c.kind_id().into(), BANG));
-
-        if !is_parens && !is_not {
-            break;
-        }
-
-        // A `!` wrapper proves the contained value is boolean even
-        // when the parent context didn't (e.g. `return !x;`).
-        if !has_boolean_content && is_not {
-            has_boolean_content = true;
-        }
-
-        // Both `parenthesized_expression` and `prefix_unary_expression`
-        // store their inner expression at child index 1.
-        let Some(child) = node.child(1) else { break };
-        node = child;
-        node_kind = node.kind_id().into();
-
-        // Found the innermost operand; count it if a boolean context
-        // was established up the chain. The `csharp_bool_terminal_kinds!()`
-        // set bundles invocation aliases, the `Identifier` /
-        // `BooleanLiteral` leaves, and the five bool-evaluating kinds
-        // restored by #372 (member access / await / cast / is-pattern /
-        // element access).
-        if matches!(node_kind, csharp_bool_terminal_kinds!()) {
-            if has_boolean_content {
-                *conditions += 1.;
-            }
-            break;
-        }
-    }
-}
-
-fn csharp_count_unary_conditions(list_node: &Node, conditions: &mut f64) {
-    use Csharp::*;
-
-    let list_kind = list_node.kind_id().into();
-    let mut cursor = list_node.cursor();
-
-    if cursor.goto_first_child() {
-        loop {
-            let node = cursor.node();
-            let node_kind = node.kind_id().into();
-
-            // `csharp_bool_terminal_kinds!()` bundles invocation aliases,
-            // `Identifier`, `BooleanLiteral`, and the bool-evaluating
-            // expression kinds restored by #372 (member access / await /
-            // cast / is-pattern / element access).
-            if matches!(node_kind, csharp_bool_terminal_kinds!())
-                && matches!(list_kind, BinaryExpression)
-            {
-                *conditions += 1.;
-            } else {
-                csharp_inspect_container(&node, conditions);
-            }
-
-            if !cursor.goto_next_sibling() {
-                break;
-            }
-        }
-    }
-}
-
-// Inspects a list of elements and counts any unary conditional expression found
-fn java_count_unary_conditions(list_node: &Node, conditions: &mut f64) {
-    use Java::*;
-
-    let list_kind = list_node.kind_id().into();
-    let mut cursor = list_node.cursor();
-
-    // Scans the immediate children nodes of the argument node
-    if cursor.goto_first_child() {
-        loop {
-            // Gets the current child node and its kind
-            let node = cursor.node();
-            let node_kind = node.kind_id().into();
-
-            // Checks if the node is a unary condition. The terminal set
-            // includes `FieldAccess`, `CastExpression`, `ArrayAccess`,
-            // and `InstanceofExpression` so that bool-evaluating
-            // operands of `&&` / `||` chains are not silently zeroed
-            // out (mirrors the C# fix in #372; lesson #19).
-            if matches!(node_kind, java_bool_terminal_kinds!())
-                && matches!(list_kind, BinaryExpression)
-            {
-                *conditions += 1.;
-            } else {
-                // Checks if the node is a unary condition container
-                java_inspect_container(&node, conditions);
-            }
-
-            // Moves the cursor to the next sibling node of the current node
-            // Exits the scan if there is no next sibling node
-            if !cursor.goto_next_sibling() {
-                break;
-            }
-        }
-    }
-}
-
-fn groovy_inspect_container(container_node: &Node, conditions: &mut f64) {
-    use Groovy::*;
-
-    let mut node = *container_node;
-    let mut node_kind = node.kind_id().into();
-
-    let Some(parent) = node.parent() else { return };
-    let mut has_boolean_content = match parent.kind_id().into() {
-        BinaryExpression | IfStatement | WhileStatement | DoWhileStatement | ForStatement => true,
-        TernaryExpression => node
-            .previous_sibling()
-            .is_none_or(|prev_node| !matches!(prev_node.kind_id().into(), QMARK | COLON)),
-        _ => false,
-    };
-
-    loop {
-        let is_parenthesised_exp = matches!(node_kind, ParenthesizedExpression);
-        let is_not_operator = matches!(node_kind, UnaryExpression)
-            && node
-                .child(0)
-                .is_some_and(|c| matches!(c.kind_id().into(), BANG));
-
-        if !is_parenthesised_exp && !is_not_operator {
-            break;
-        }
-
-        if !has_boolean_content && is_not_operator {
-            has_boolean_content = true;
-        }
-
-        let Some(child) = node.child(1) else { break };
-        node = child;
-        node_kind = node.kind_id().into();
-
-        // `BooleanLiteral` is the dekobon tree-sitter-groovy
-        // grammar's named wrapper for `true` / `false` — see the
-        // doc comment on `groovy_count_condition`. The remaining
-        // bool-evaluating terminals (`FieldAccess`, `CastExpression`,
-        // `ParenthesizedTypeCast`, `InstanceofExpression`) mirror
-        // the C# fix in #372 (lesson #19).
-        if matches!(node_kind, groovy_bool_terminal_kinds!()) {
-            if has_boolean_content {
-                *conditions += 1.;
-            }
-            break;
-        }
-    }
-}
-
-fn groovy_count_unary_conditions(list_node: &Node, conditions: &mut f64) {
-    use Groovy::*;
-
-    let list_kind = list_node.kind_id().into();
-    let mut cursor = list_node.cursor();
-
-    if cursor.goto_first_child() {
-        loop {
-            let node = cursor.node();
-            let node_kind = node.kind_id().into();
-
-            // Terminal set mirrors `groovy_inspect_container` —
-            // bool-evaluating kinds (`FieldAccess`, `CastExpression`,
-            // `ParenthesizedTypeCast`, `InstanceofExpression`) added
-            // per issue #372 (lesson #19).
-            if matches!(node_kind, groovy_bool_terminal_kinds!())
-                && matches!(list_kind, BinaryExpression)
-            {
-                *conditions += 1.;
-            } else {
-                groovy_inspect_container(&node, conditions);
-            }
-
             if !cursor.goto_next_sibling() {
                 break;
             }
@@ -665,3746 +408,6 @@ fn groovy_count_unary_conditions(list_node: &Node, conditions: &mut f64) {
 //   - PreprocCode, CcommentCode: no executable code (comments /
 //     preprocessor lines only).
 implement_metric_trait!(Abc, PreprocCode, CcommentCode);
-
-// JS / TS / TSX / Mozjs share an expression / statement vocabulary;
-// the helper macro below generates the per-language unary-conditional
-// walker pair (Fitzpatrick Rule 9 / Listing 2; issue #403). Each
-// `&&` / `||` token in the dispatcher routes through
-// `<lang>_count_unary_conditions` and counts the immediate operands
-// of the parent `binary_expression` once. Operands wrapped in `(…)` /
-// `!…` are unwrapped via `<lang>_inspect_container`. Terminal-bool
-// kinds include `Identifier`, the boolean literal tokens `True` /
-// `False`, plus `CallExpression` / `NewExpression` (object
-// construction in JS / TS) / `MemberExpression` / `SubscriptExpression`
-// — every expression kind whose evaluated value is implicitly boolean
-// in an `if` / `while` / ternary slot.
-macro_rules! impl_js_family_unary_walker {
-    ($Lang:ident, $inspect:ident, $count:ident, $terminals:path) => {
-        fn $inspect(container_node: &Node, conditions: &mut f64) {
-            use $Lang::*;
-
-            let mut node = *container_node;
-            let mut node_kind = node.kind_id().into();
-            let Some(parent) = node.parent() else { return };
-            let parent_kind = parent.kind_id().into();
-            let mut has_boolean_content = matches!(
-                parent_kind,
-                BinaryExpression | IfStatement | WhileStatement | DoStatement | ForStatement
-            ) || (matches!(parent_kind, TernaryExpression)
-                && node
-                    .previous_sibling()
-                    .is_none_or(|prev| !matches!(prev.kind_id().into(), QMARK | COLON)));
-
-            loop {
-                let is_parens = matches!(node_kind, ParenthesizedExpression);
-                let is_not = matches!(node_kind, UnaryExpression)
-                    && node.child(0).is_some_and(|c| c.kind_id() == BANG as u16);
-
-                if !is_parens && !is_not {
-                    break;
-                }
-                if !has_boolean_content && is_not {
-                    has_boolean_content = true;
-                }
-
-                let Some(child) = node.child(1) else { break };
-                node = child;
-                node_kind = node.kind_id().into();
-
-                if matches!(node_kind, $terminals!()) {
-                    if has_boolean_content {
-                        *conditions += 1.;
-                    }
-                    break;
-                }
-            }
-        }
-
-        fn $count(list_node: &Node, conditions: &mut f64) {
-            use $Lang::*;
-
-            let list_kind = list_node.kind_id().into();
-            let mut cursor = list_node.cursor();
-
-            if cursor.goto_first_child() {
-                loop {
-                    let node = cursor.node();
-                    let node_kind = node.kind_id().into();
-
-                    if matches!(node_kind, $terminals!()) && matches!(list_kind, BinaryExpression) {
-                        *conditions += 1.;
-                    } else if node.is_named() {
-                        $inspect(&node, conditions);
-                    }
-
-                    if !cursor.goto_next_sibling() {
-                        break;
-                    }
-                }
-            }
-        }
-    };
-}
-
-impl_js_family_unary_walker!(
-    Typescript,
-    typescript_inspect_container,
-    typescript_count_unary_conditions,
-    typescript_bool_terminal_kinds
-);
-impl_js_family_unary_walker!(
-    Tsx,
-    tsx_inspect_container,
-    tsx_count_unary_conditions,
-    tsx_bool_terminal_kinds
-);
-impl_js_family_unary_walker!(
-    Javascript,
-    javascript_inspect_container,
-    javascript_count_unary_conditions,
-    javascript_bool_terminal_kinds
-);
-impl_js_family_unary_walker!(
-    Mozjs,
-    mozjs_inspect_container,
-    mozjs_count_unary_conditions,
-    mozjs_bool_terminal_kinds
-);
-
-// TypeScript / TSX share the same expression / statement vocabulary;
-// the `ts_abc_compute!` macro expands the same token-level
-// Fitzpatrick rules for both. Conditions capture every comparison and
-// control-flow arm (the original token-level set), plus Phase-2 walker
-// arms for `&&` / `||` operand counting and the
-// `IfStatement` / `WhileStatement` / `DoStatement` / `ReturnStatement`
-// / `Arguments` slots — each of those arms routes through the
-// language's `$inspect_container` (paren / unary unwrap) and
-// `$count_unary` (operand walker) helpers generated by
-// `impl_js_family_unary_walker!`.
-//
-// Declaration sentinel: `lexical_declaration` and `variable_declaration`
-// push a `Var` sentinel that suppresses counting the initializer `=` as
-// an assignment. The `Const` token promotes to `Const` (compile-time
-// constant — initializer is not a mutable assignment). `let` and `var`
-// keep the `Var` slot. Augmented assignments (`+=`) and update
-// expressions (`++`, `--`) always count.
-macro_rules! ts_abc_compute {
-    ($lang:ident, $count_unary:path, $inspect_container:path) => {
-        fn compute<'a>(node: &Node<'a>, _code: &'a [u8], stats: &mut Stats) {
-            use $lang::*;
-
-            match node.kind_id().into() {
-                // Augmented assignments and pre/post increment/decrement
-                // always count.
-                PLUSEQ | DASHEQ | STAREQ | SLASHEQ | PERCENTEQ | STARSTAREQ | AMPEQ | PIPEEQ
-                | CARETEQ | LTLTEQ | GTGTEQ | GTGTGTEQ | AMPAMPEQ | PIPEPIPEEQ | QMARKQMARKEQ
-                | PLUSPLUS | DASHDASH => {
-                    stats.assignments += 1.;
-                }
-                // Variable declarations push a `Var` sentinel; the `Const`
-                // keyword promotes the top to `Const` so the initializer
-                // `=` is treated as a constant binding.
-                LexicalDeclaration | VariableDeclaration => {
-                    stats.declaration.push(DeclKind::Var);
-                }
-                Const => {
-                    if let Some(DeclKind::Var) = stats.declaration.last() {
-                        stats.declaration.push(DeclKind::Const);
-                    }
-                }
-                SEMI => {
-                    if let Some(DeclKind::Const | DeclKind::Var) = stats.declaration.last() {
-                        stats.declaration.clear();
-                    }
-                }
-                // Plain `=` outside `const` declarations is an assignment.
-                EQ if !matches!(stats.declaration.last(), Some(DeclKind::Const)) => {
-                    stats.assignments += 1.;
-                }
-                // Function invocation and object construction count as
-                // branches. Member calls and chained calls all surface
-                // as `CallExpression`.
-                CallExpression | NewExpression => {
-                    stats.branches += 1.;
-                }
-                // Comparison and equality operators, ternary `?`, `??`,
-                // `instanceof`, `else`, `case`, `catch`, `try`. The
-                // `default` arm of a `switch` is intentionally NOT a
-                // condition: it is the unconditional fallthrough, so
-                // cyclomatic counts only the `Case` arms (issue #469).
-                // Both the statement (`default:`) and arrow
-                // (`default ->`) forms emit the same `Default` token, so
-                // omitting it here covers both.
-                EQEQ | EQEQEQ | BANGEQ | BANGEQEQ | LTEQ | GTEQ | QMARK | QMARKQMARK
-                | Instanceof | Else | Case | Try | Catch => {
-                    stats.conditions += 1.;
-                }
-                // `<` and `>` may also delimit type arguments / type
-                // parameters (`Array<number>`, `class Foo<T> {}`); skip
-                // those, count only comparison usage.
-                GT | LT
-                    if node.parent().is_some_and(|p| {
-                        !matches!(p.kind_id().into(), TypeArguments | TypeParameters)
-                    }) =>
-                {
-                    stats.conditions += 1.;
-                }
-                // Fitzpatrick Rule 9: each operand of a `&&` / `||`
-                // chain is one condition (issue #403).
-                AMPAMP | PIPEPIPE => {
-                    if let Some(parent) = node.parent() {
-                        $count_unary(&parent, &mut stats.conditions);
-                    }
-                }
-                // Phase-2B (issue #403): condition slots. JS / TS
-                // wrap `if (...)` / `while (...)` / `do {…} while
-                // (...)` in `parenthesized_expression`, so
-                // `<lang>_inspect_container`'s paren-unwrap handles
-                // the boolean-literal case (`if (true)` counts 1).
-                // The condition sits at child(1) for if and while.
-                // For `do_statement`, the condition is at child(3)
-                // (children: `do`(0), body(1), `while`(2),
-                // parenthesized condition(3), `;`(4)).
-                IfStatement | WhileStatement => {
-                    if let Some(cond) = node.child(1) {
-                        $inspect_container(&cond, &mut stats.conditions);
-                    }
-                }
-                DoStatement => {
-                    // children: `do`(0), body(1), `while`(2),
-                    // parenthesized condition(3), `;`(4).
-                    if let Some(cond) = node.child(3) {
-                        $inspect_container(&cond, &mut stats.conditions);
-                    }
-                }
-                // `return value;` — value at child(1). The bare
-                // `return;` (no value) form has no child(1).
-                ReturnStatement => {
-                    if let Some(value) = node.child(1) {
-                        $inspect_container(&value, &mut stats.conditions);
-                    }
-                }
-                // Method-argument walker for `f(!a, !b)`.
-                Arguments => {
-                    $count_unary(node, &mut stats.conditions);
-                }
-                _ => {}
-            }
-        }
-    };
-}
-
-impl Abc for TypescriptCode {
-    ts_abc_compute!(
-        Typescript,
-        typescript_count_unary_conditions,
-        typescript_inspect_container
-    );
-}
-
-impl Abc for TsxCode {
-    ts_abc_compute!(Tsx, tsx_count_unary_conditions, tsx_inspect_container);
-}
-
-// JavaScript / Mozjs share TypeScript's expression / statement
-// vocabulary. The `js_abc_compute!` macro expands the same
-// token-level Fitzpatrick rules as `ts_abc_compute!`, with two
-// adjustments:
-//
-//   1. `LT` / `GT` are always comparison operators in plain JS — there
-//      are no `TypeArguments` / `TypeParameters` nodes to gate against.
-//   2. JS retains the same `LexicalDeclaration` / `VariableDeclaration`
-//      sentinel handling so `const x = 5` does not double-count the
-//      initializer `=` as an assignment. `let x = 5` and `var x = 5`
-//      DO count their initializer `=` as an assignment — only `const`
-//      suppresses, matching the TS impl above. This deliberately
-//      deviates from a strict reading of Fitzpatrick's "declaration
-//      initialiser is not an assignment" rule because `let`/`var`
-//      bindings can be reassigned and the initial value is the first
-//      assignment of the binding's lifetime.
-macro_rules! js_abc_compute {
-    ($lang:ident, $count_unary:path, $inspect_container:path) => {
-        fn compute<'a>(node: &Node<'a>, _code: &'a [u8], stats: &mut Stats) {
-            use $lang::*;
-
-            match node.kind_id().into() {
-                PLUSEQ | DASHEQ | STAREQ | SLASHEQ | PERCENTEQ | STARSTAREQ | AMPEQ | PIPEEQ
-                | CARETEQ | LTLTEQ | GTGTEQ | GTGTGTEQ | AMPAMPEQ | PIPEPIPEEQ | QMARKQMARKEQ
-                | PLUSPLUS | DASHDASH => {
-                    stats.assignments += 1.;
-                }
-                LexicalDeclaration | VariableDeclaration => {
-                    stats.declaration.push(DeclKind::Var);
-                }
-                Const => {
-                    if let Some(DeclKind::Var) = stats.declaration.last() {
-                        stats.declaration.push(DeclKind::Const);
-                    }
-                }
-                SEMI => {
-                    if let Some(DeclKind::Const | DeclKind::Var) = stats.declaration.last() {
-                        stats.declaration.clear();
-                    }
-                }
-                EQ if !matches!(stats.declaration.last(), Some(DeclKind::Const)) => {
-                    stats.assignments += 1.;
-                }
-                CallExpression | NewExpression => {
-                    stats.branches += 1.;
-                }
-                // The `default` arm is the unconditional fallthrough and
-                // is excluded, mirroring cyclomatic's `Case`-only count
-                // (issue #469); see the TS macro above for the rationale.
-                EQEQ | EQEQEQ | BANGEQ | BANGEQEQ | LTEQ | GTEQ | LT | GT | QMARK | QMARKQMARK
-                | Instanceof | Else | Case | Try | Catch => {
-                    stats.conditions += 1.;
-                }
-                // Fitzpatrick Rule 9: each operand of a `&&` / `||`
-                // chain is one condition (issue #403).
-                AMPAMP | PIPEPIPE => {
-                    if let Some(parent) = node.parent() {
-                        $count_unary(&parent, &mut stats.conditions);
-                    }
-                }
-                // Phase-2B (issue #403): condition slots. Same shape
-                // as the TypeScript impl above — see that macro's
-                // arm-block for the per-child-index rationale.
-                IfStatement | WhileStatement => {
-                    if let Some(cond) = node.child(1) {
-                        $inspect_container(&cond, &mut stats.conditions);
-                    }
-                }
-                DoStatement => {
-                    // children: `do`(0), body(1), `while`(2),
-                    // parenthesized condition(3), `;`(4).
-                    if let Some(cond) = node.child(3) {
-                        $inspect_container(&cond, &mut stats.conditions);
-                    }
-                }
-                ReturnStatement => {
-                    if let Some(value) = node.child(1) {
-                        $inspect_container(&value, &mut stats.conditions);
-                    }
-                }
-                Arguments => {
-                    $count_unary(node, &mut stats.conditions);
-                }
-                _ => {}
-            }
-        }
-    };
-}
-
-impl Abc for JavascriptCode {
-    js_abc_compute!(
-        Javascript,
-        javascript_count_unary_conditions,
-        javascript_inspect_container
-    );
-}
-
-impl Abc for MozjsCode {
-    js_abc_compute!(Mozjs, mozjs_count_unary_conditions, mozjs_inspect_container);
-}
-
-// Fitzpatrick's ABC rules adapted for Kotlin syntax. Kotlin shares the
-// JVM and Java's spec roots: assignments count once per `=` / augmented
-// assignment / ++ / --, branches count once per function invocation or
-// object construction, conditions count comparison operators plus the
-// `else` / `when`-entry / `catch` arms. Compared with the Java impl we
-// stay token-level (matching the leaf kind_ids) rather than walking
-// `Modifiers` children; the Kotlin grammar exposes the relevant
-// operators directly as token nodes inside `binary_expression`,
-// `assignment`, `prefix_expression`, and `postfix_expression`.
-
-// Returns true when this `=` token initialises an *immutable* (`val`)
-// binding, whose initialiser is part of the declaration and therefore not
-// an ABC assignment. The decision is structural: the `=` must be a direct
-// child of a `property_declaration` or `class_parameter`, and that parent
-// must carry a `val` keyword child. A `var`/plain declaration initialiser
-// and any standalone `assignment` return false (they count).
-fn kotlin_eq_initializes_immutable_binding(eq_node: &Node) -> bool {
-    use Kotlin::*;
-
-    let Some(parent) = eq_node.parent() else {
-        return false;
-    };
-    if !matches!(
-        parent.kind_id().into(),
-        PropertyDeclaration | ClassParameter
-    ) {
-        return false;
-    }
-    parent.children().any(|child| child.kind_id() == Val)
-}
-
-// Kotlin ABC unary-conditional walker (Fitzpatrick Rule 9; issue #557).
-// tree-sitter-kotlin-ng parses `a && b || c` as a left-nested chain of
-// flat `binary_expression` nodes carrying `&&` / `||` operator tokens,
-// the same shape as the Java template. Negation surfaces as
-// `unary_expression` whose child(0) is the `!` token; the condition slot
-// may also be wrapped in `parenthesized_expression`. Both are unwrapped
-// by `kotlin_inspect_container` to reach the inner bare operand.
-fn kotlin_inspect_container(container_node: &Node, conditions: &mut f64) {
-    use Kotlin::*;
-
-    let mut node = *container_node;
-    let mut node_kind = node.kind_id().into();
-    let Some(parent) = node.parent() else { return };
-    // A parenthesised / negated operand only contributes when it sits in
-    // a boolean-evaluating slot. The chain wrapper (`binary_expression`)
-    // and the control-flow headers all qualify; a `!`-operator anywhere
-    // also proves the operand is boolean (`if (!x)`).
-    let mut has_boolean_content = matches!(
-        parent.kind_id().into(),
-        BinaryExpression | IfExpression | WhileStatement | DoWhileStatement | ForStatement
-    );
-
-    loop {
-        let is_parens = matches!(node_kind, ParenthesizedExpression);
-        let is_not = matches!(node_kind, UnaryExpression)
-            && node.child(0).is_some_and(|c| c.kind_id() == BANG as u16);
-
-        if !is_parens && !is_not {
-            break;
-        }
-        if !has_boolean_content && is_not {
-            has_boolean_content = true;
-        }
-
-        // Parenthesised expressions wrap their inner expression at child
-        // index 1 (after the `(` token); a `!` unary stores its operand
-        // at index 1 (after the `!` token).
-        let Some(child) = node.child(1) else { break };
-        node = child;
-        node_kind = node.kind_id().into();
-
-        if matches!(node_kind, kotlin_bool_terminal_kinds!()) {
-            if has_boolean_content {
-                *conditions += 1.;
-            }
-            break;
-        }
-    }
-}
-
-// Counts each non-comparison operand of a Kotlin `&&` / `||` chain once.
-// Mirrors `java_count_unary_conditions`: comparison operands are nested
-// `binary_expression` nodes (absent from `kotlin_bool_terminal_kinds!()`)
-// and so contribute nothing, while bare identifiers / calls / member
-// accesses each add one. Inner chain links and `!` / paren wrappers are
-// routed through `kotlin_inspect_container`.
-fn kotlin_count_unary_conditions(list_node: &Node, conditions: &mut f64) {
-    use Kotlin::*;
-
-    let list_kind = list_node.kind_id().into();
-    let mut cursor = list_node.cursor();
-
-    if cursor.goto_first_child() {
-        loop {
-            let node = cursor.node();
-            let node_kind = node.kind_id().into();
-
-            if matches!(node_kind, kotlin_bool_terminal_kinds!())
-                && matches!(list_kind, BinaryExpression)
-            {
-                *conditions += 1.;
-            } else if node.is_named() {
-                kotlin_inspect_container(&node, conditions);
-            }
-
-            if !cursor.goto_next_sibling() {
-                break;
-            }
-        }
-    }
-}
-
-// Count the bare-predicate condition of a Kotlin `if`/`while`/`do-while`
-// as one condition — Fitzpatrick's "unary conditional expression",
-// mirroring `ruby_count_condition` / `rust_count_condition`.
-// tree-sitter-kotlin-ng exposes the predicate via the `condition` field on
-// `if_expression`, `while_statement`, and `do_while_statement`, so the
-// field lookup is position-independent across all three forms. A bare
-// terminal (`if (flag)`) counts directly; a comparison or boolean chain
-// (`if (a == b)`, `if (a && b)`) is a nested `binary_expression` already
-// counted by the comparison-token and `&&`/`||` walker arms, so it adds
-// nothing here. A parenthesised or negated predicate (`if ((flag))`,
-// `if (!flag)`) is unwrapped by `kotlin_inspect_container`. Without this
-// arm, idiomatic Kotlin bare predicates reported 0 ABC conditions while
-// Kotlin's own cyclomatic counted the decision, breaking the
-// conditions >= decisions invariant (#469/#473/#456/#696); issue #773.
-fn kotlin_count_condition(condition: &Node, conditions: &mut f64) {
-    use Kotlin::*;
-    let kind = condition.kind_id().into();
-    if matches!(kind, kotlin_bool_terminal_kinds!()) {
-        *conditions += 1.;
-    } else if matches!(kind, ParenthesizedExpression | UnaryExpression) {
-        kotlin_inspect_container(condition, conditions);
-    }
-}
-
-impl Abc for KotlinCode {
-    fn compute<'a>(node: &Node<'a>, _code: &'a [u8], stats: &mut Stats) {
-        use Kotlin::*;
-
-        match node.kind_id().into() {
-            // Augmented assignments and pre/post increment-decrement
-            // always count, regardless of declaration context.
-            PLUSEQ | DASHEQ | STAREQ | SLASHEQ | PERCENTEQ | PLUSPLUS | DASHDASH => {
-                stats.assignments += 1.;
-            }
-            // Plain `=` token. A declaration initialiser (`val`/`var x = …`,
-            // primary-constructor parameter default `class C(val a = 5)`) is
-            // "part of the declaration" (Fitzpatrick), so the `=` is counted
-            // only for *mutable* bindings; an immutable `val` initialiser is
-            // suppressed. A standalone `assignment` always counts.
-            //
-            // This is decided structurally from the `=` token's parent node
-            // rather than a persistent declaration stack. tree-sitter-kotlin
-            // does NOT emit a `SEMI` token even for explicit semicolons, and
-            // newline-terminated statements emit no terminator at all, so a
-            // stack cleared on `SEMI` (the pre-#455 design) never cleared:
-            // the immutable-`val` sentinel leaked and suppressed every later
-            // standalone assignment in the same function (issue #455).
-            EQ if !kotlin_eq_initializes_immutable_binding(node) => {
-                stats.assignments += 1.;
-            }
-            // Branches: every call expression plus object construction.
-            // Kotlin's `new` is implicit — `Foo()` parses as
-            // `CallExpression` with a type-named receiver. The
-            // Halstead-side classification treats it uniformly. Indexed
-            // access (`arr[i]`) is NOT a branch (it's an operator on a
-            // sequence), matching the Java rule of "method invocation
-            // only".
-            CallExpression => {
-                stats.branches += 1.;
-            }
-            // Conditions: comparison operators, identity equality,
-            // ternary-elvis (`?:`), `as?` safe-cast, and the arms of
-            // control-flow constructs (`else`, `catch`, `when` entries).
-            // Kotlin's `if`-expression does not need an extra count for
-            // the `if` keyword itself — Fitzpatrick counts the
-            // *conditions*, and the unary condition is already implicit
-            // in the boolean operand. We add the `if` arm via the `Else`
-            // keyword for else-branches and via `WhenEntry` for `when`.
-            // `Try` is the `try` keyword token of a `try_expression`.
-            // Fitzpatrick counts both `try` and `catch` as conditions, and
-            // Java / C# / C++ / Groovy already count both; Kotlin previously
-            // counted only `CatchBlock`, so `try {} catch (e) {}` scored one
-            // fewer condition here than in every sibling (#696).
-            LTEQ | GTEQ | EQEQ | EQEQEQ | BANGEQ | BANGEQEQ | Try | CatchBlock | QMARKCOLON
-            | AsQMARK => {
-                stats.conditions += 1.;
-            }
-            // Phase-2B condition slot: the bare predicate of an
-            // `if`/`while`/`do-while` is one unary condition. The
-            // `condition` field locates the predicate position-
-            // independently across all three forms. `kotlin_count_condition`
-            // counts a bare terminal directly and routes paren/negation
-            // wrappers through `kotlin_inspect_container`, while a comparison
-            // or `&&`/`||` predicate is a `binary_expression` already counted
-            // by the token arms above — so no double-count (#773).
-            IfExpression | WhileStatement | DoWhileStatement => {
-                if let Some(condition) = node.child_by_field_name("condition") {
-                    kotlin_count_condition(&condition, &mut stats.conditions);
-                }
-            }
-            // A `when` entry is a decision point except for the `else ->`
-            // fallback arm, which is the analogue of C-family `default:`
-            // and Rust's wildcard `_ =>`. Cyclomatic already excludes it
-            // (`WhenEntry if !kotlin_when_entry_is_else`); ABC must track
-            // the same decision count (issue #456, lesson 11).
-            WhenEntry if !crate::metrics::cyclomatic::kotlin_when_entry_is_else(node) => {
-                stats.conditions += 1.;
-            }
-            // `else` is a keyword token used in both `if_expression`'s
-            // else-clause and `when`'s `else ->` entry. Only count it
-            // when it belongs to an `if_expression`; the `WhenEntry`
-            // wrapper above already covers the `when` case.
-            Else if node.parent().is_some_and(|p| p.kind_id() == IfExpression) => {
-                stats.conditions += 1.;
-            }
-            // `<` and `>` may appear as type-argument brackets
-            // (`List<Int>`); exclude those by checking the parent kind.
-            LT | GT
-                if node.parent().is_some_and(|p| {
-                    !matches!(p.kind_id().into(), TypeArguments | TypeParameters)
-                }) =>
-            {
-                stats.conditions += 1.;
-            }
-            // Fitzpatrick Rule 9 walker: each non-comparison operand of a
-            // `&&` / `||` chain is one condition (issue #557). The short-
-            // circuit operators are not counted directly (cross-language
-            // policy, #395); the walker fires off the operator token and
-            // inspects the parent `binary_expression`.
-            AMPAMP | PIPEPIPE => {
-                if let Some(parent) = node.parent() {
-                    kotlin_count_unary_conditions(&parent, &mut stats.conditions);
-                }
-            }
-            _ => {}
-        }
-    }
-}
-
-// PHP ABC unary-conditional walker (Fitzpatrick Rule 9; issue #403).
-// PHP's grammar uses `unary_op_expression` (not `unary_expression`) for
-// `!` and `~` prefix operators. Terminal-bool kinds: `Name` (function/
-// constant identifier — the bare-identifier kind in tree-sitter-php),
-// `VariableName` (`$x`), `Boolean` (the named `true` / `false` wrapper),
-// and every call / member-access / subscript form. `ParenthesizedExpression`
-// wraps `if (...)`-style condition slots.
-fn php_inspect_container(container_node: &Node, conditions: &mut f64) {
-    use Php::*;
-
-    let mut node = *container_node;
-    let mut node_kind = node.kind_id().into();
-    let Some(parent) = node.parent() else { return };
-    let parent_kind = parent.kind_id().into();
-    let mut has_boolean_content = matches!(
-        parent_kind,
-        BinaryExpression | IfStatement | WhileStatement | DoStatement | ForStatement
-    ) || (matches!(parent_kind, ConditionalExpression)
-        && node
-            .previous_sibling()
-            .is_none_or(|prev| !matches!(prev.kind_id().into(), QMARK | COLON)));
-
-    loop {
-        let is_parens = matches!(node_kind, ParenthesizedExpression);
-        let is_not = matches!(node_kind, UnaryOpExpression)
-            && node.child(0).is_some_and(|c| c.kind_id() == BANG as u16);
-
-        if !is_parens && !is_not {
-            break;
-        }
-        if !has_boolean_content && is_not {
-            has_boolean_content = true;
-        }
-
-        let Some(child) = node.child(1) else { break };
-        node = child;
-        node_kind = node.kind_id().into();
-
-        if matches!(node_kind, php_bool_terminal_kinds!()) {
-            if has_boolean_content {
-                *conditions += 1.;
-            }
-            break;
-        }
-    }
-}
-
-// Phase-2B helper (issue #403): pass `node.child(idx)` through
-// `php_inspect_container`. PHP wraps `if (...)` / `while (...)` /
-// `do {…} while (...)` in `parenthesized_expression`, so the paren
-// unwrap handles the boolean-literal case (`if (true)` counts 1).
-fn php_inspect_child(node: &Node, idx: usize, conditions: &mut f64) {
-    if let Some(child) = node.child(idx) {
-        php_inspect_container(&child, conditions);
-    }
-}
-
-// Returns the value slot of a PHP `argument` wrapper node.
-// Positional argument `m(!$a)` has a single named child — the value.
-// Named argument `m(name: !$a)` has children `name`, `:`, value — the
-// last named child is the value. Returns the last named child for
-// both shapes; returns None only when the argument has no named
-// children (grammar-error case).
-fn php_argument_value<'a>(argument: &Node<'a>) -> Option<Node<'a>> {
-    let mut cursor = argument.cursor();
-    let mut last_named = None;
-    if cursor.goto_first_child() {
-        loop {
-            let child = cursor.node();
-            if child.is_named() {
-                last_named = Some(child);
-            }
-            if !cursor.goto_next_sibling() {
-                break;
-            }
-        }
-    }
-    last_named
-}
-
-fn php_count_unary_conditions(list_node: &Node, conditions: &mut f64) {
-    use Php::*;
-
-    let list_kind = list_node.kind_id().into();
-    let mut cursor = list_node.cursor();
-
-    if cursor.goto_first_child() {
-        loop {
-            let node = cursor.node();
-            let node_kind = node.kind_id().into();
-
-            // PHP wraps each call argument in an `argument` node;
-            // descend through that wrapper to the value slot. For
-            // named arguments `m(name: !$a)` the value is the LAST
-            // named child (`name`/`:`/`value`); for positional
-            // arguments `m(!$a)` the value is the only child. Use
-            // the last named child to handle both shapes — and skip
-            // the rare grammar-error case where Argument has no
-            // named children.
-            let inner = if matches!(node_kind, Argument) {
-                let Some(value) = php_argument_value(&node) else {
-                    if !cursor.goto_next_sibling() {
-                        break;
-                    }
-                    continue;
-                };
-                value
-            } else {
-                node
-            };
-            let inner_kind = inner.kind_id().into();
-
-            if matches!(inner_kind, php_bool_terminal_kinds!())
-                && matches!(list_kind, BinaryExpression)
-            {
-                *conditions += 1.;
-            } else if inner.is_named() {
-                php_inspect_container(&inner, conditions);
-            }
-
-            if !cursor.goto_next_sibling() {
-                break;
-            }
-        }
-    }
-}
-
-impl Abc for PhpCode {
-    fn compute<'a>(node: &Node<'a>, _code: &'a [u8], stats: &mut Stats) {
-        use Php::*;
-
-        match node.kind_id().into() {
-            // Assignments: explicit assignment expressions and augmented forms,
-            // plus pre/post increment and decrement. `const_declaration` and
-            // `enum_case` use their own `const_element` / value-assignment
-            // shapes, so they do not produce `AssignmentExpression` nodes —
-            // matching the assignment-expression kinds naturally excludes
-            // them.
-            AssignmentExpression
-            | AugmentedAssignmentExpression
-            | ReferenceAssignmentExpression
-            | PLUSPLUS
-            | DASHDASH => {
-                stats.assignments += 1.;
-            }
-            // Branches: every PHP call kind plus object construction.
-            FunctionCallExpression
-            | MemberCallExpression
-            | ScopedCallExpression
-            | NullsafeMemberCallExpression
-            | ObjectCreationExpression => {
-                stats.branches += 1.;
-            }
-            // Conditions: comparison and identity operators (anonymous tokens
-            // inside `binary_expression`), `instanceof`, ternary `?`, and
-            // control-flow arms.
-            EQEQ
-            | EQEQEQ
-            | BANGEQ
-            | BANGEQEQ
-            | LT
-            | GT
-            | LTEQ
-            | GTEQ
-            | LTEQGT
-            | LTGT
-            | Instanceof
-            | ConditionalExpression
-            | ElseClause
-            | ElseClause2
-            | ElseIfClause
-            | ElseIfClause2
-            // `case` arms and non-default `match` arms are conditions;
-            // the `default:` (`DefaultStatement`) and `default =>`
-            // (`MatchDefaultExpression`) arms are NOT — they are the
-            // unconditional fallthrough, which PHP's cyclomatic gate also
-            // excludes (it counts `CaseStatement | MatchConditionalExpression`
-            // only). Dropping both Default kinds keeps ABC conditions equal
-            // to the cyclomatic decision count (issue #473, mirroring the
-            // #469 C-family fix and #456 Kotlin/C# fixes).
-            | CaseStatement
-            | MatchConditionalExpression
-            | CatchClause => {
-                stats.conditions += 1.;
-            }
-            // Fitzpatrick Rule 9: each operand of a `&&` / `||` / `and`
-            // / `or` / `xor` chain is one condition (issue #403). PHP
-            // exposes both the punctuation forms (`&&`, `||`) and the
-            // low-precedence keyword forms (`and`, `or`, `xor`) as
-            // distinct tokens inside `binary_expression`; both fire
-            // the walker so `connect() or die();`-style idiom counts
-            // the same as `connect() || die();`.
-            AMPAMP | PIPEPIPE | And | Or | Xor => {
-                if let Some(parent) = node.parent() {
-                    php_count_unary_conditions(&parent, &mut stats.conditions);
-                }
-            }
-            // Phase-2B (issue #403): condition slots. PHP wraps
-            // `if (...)` / `while (...)` in `parenthesized_expression`
-            // at child(1); `return value;` exposes the value at the
-            // same index. `do {…} while (...)` has the parenthesized
-            // condition at child(3).
-            IfStatement | WhileStatement | ReturnStatement => {
-                php_inspect_child(node, 1, &mut stats.conditions);
-            }
-            DoStatement => {
-                php_inspect_child(node, 3, &mut stats.conditions);
-            }
-            // `f(!$a, !$b)` — argument list walker.
-            Arguments => {
-                php_count_unary_conditions(node, &mut stats.conditions);
-            }
-            _ => {}
-        }
-    }
-}
-
-// Ruby ABC rules follow the Fitzpatrick paper's spirit, adapted to
-// tree-sitter-ruby:
-// - Assignments: `assignment` (plain `=`) and `operator_assignment`
-//   (`+=`, `-=`, `||=`, `&&=`, …). Tree-sitter wraps both forms in a
-//   dedicated node, so we count one assignment per node and avoid
-//   double-counting the inner `=` / augmented token.
-// - Branches: every Ruby method invocation kind (`Call` / `Call2` /
-//   `Call3` / `Call4`) plus `super` and `yield`. `yield` is grammar-
-//   level a "block invocation" but ABC's branch bucket is "message
-//   pass / function call", so it belongs here. `attr_*` macros are
-//   `Call3` nodes and are counted as branches like any other call.
-// - Conditions: comparison and equality operator tokens emitted inside
-//   `binary` (`==`, `!=`, `===`, `<`, `>`, `<=`, `>=`, `<=>`,
-//   `=~`, `!~`), plus the control-flow arms that the Fitzpatrick rules
-//   list — the named clause nodes `Else` / `Elsif` / `When` and the
-//   `?` ternary marker, plus `Rescue` (the rescue clause) and rescue
-//   modifiers. `if` / `unless` themselves are not counted (the head
-//   condition appears as the inner comparison); the `Then` clause is
-//   an implicit grammar wrapper around every `if` / `elsif` body and
-//   is NOT counted as a separate arm.
-// Ruby ABC unary-conditional walker (Fitzpatrick Rule 9; issue #557).
-// tree-sitter-ruby parses `a && b || c` as a left-nested chain of
-// `binary` nodes carrying `&&` / `||` / `and` / `or` operator tokens
-// (the `binary` kind is aliased `Binary`..`Binary3` per lesson #2, so
-// every alias must be matched). Negation surfaces as `unary`
-// (`Unary`..`Unary5`) whose child(0) is the `!` token; the condition
-// slot may be wrapped in `parenthesized_statements`. Both are unwrapped
-// by `ruby_inspect_container`.
-fn ruby_inspect_container(container_node: &Node, conditions: &mut f64) {
-    use Ruby::*;
-
-    let mut node = *container_node;
-    let mut node_kind = node.kind_id().into();
-    let Some(parent) = node.parent() else { return };
-    let mut has_boolean_content = matches!(
-        parent.kind_id().into(),
-        Binary | Binary2 | Binary3 | If | Unless | While | Until
-    );
-
-    loop {
-        let is_parens = matches!(node_kind, ParenthesizedStatements);
-        let is_not = matches!(node_kind, Unary | Unary2 | Unary3 | Unary4 | Unary5)
-            && node.child(0).is_some_and(|c| c.kind_id() == BANG as u16);
-
-        if !is_parens && !is_not {
-            break;
-        }
-        if !has_boolean_content && is_not {
-            has_boolean_content = true;
-        }
-
-        // A `!` unary stores its operand at child index 1 (after the `!`
-        // token). `parenthesized_statements` wraps its body in named
-        // children; descend through the first named child carrying the
-        // expression.
-        let next = if is_not {
-            node.child(1)
-        } else {
-            node.children().find(Node::is_named)
-        };
-        let Some(child) = next else { break };
-        node = child;
-        node_kind = node.kind_id().into();
-
-        if matches!(node_kind, ruby_bool_terminal_kinds!()) {
-            if has_boolean_content {
-                *conditions += 1.;
-            }
-            break;
-        }
-    }
-}
-
-// Counts each non-comparison operand of a Ruby `&&` / `||` chain once.
-// Comparison operands are nested `binary` nodes (absent from
-// `ruby_bool_terminal_kinds!()`) and so contribute nothing.
-fn ruby_count_unary_conditions(list_node: &Node, conditions: &mut f64) {
-    use Ruby::*;
-
-    let list_kind = list_node.kind_id().into();
-    let mut cursor = list_node.cursor();
-
-    if cursor.goto_first_child() {
-        loop {
-            let node = cursor.node();
-            let node_kind = node.kind_id().into();
-
-            if matches!(node_kind, ruby_bool_terminal_kinds!())
-                && matches!(list_kind, Binary | Binary2 | Binary3)
-            {
-                *conditions += 1.;
-            } else if node.is_named() {
-                ruby_inspect_container(&node, conditions);
-            }
-
-            if !cursor.goto_next_sibling() {
-                break;
-            }
-        }
-    }
-}
-
-// Count the bare-predicate condition of a Ruby `if`/`unless`/`while`/
-// `until` (block or modifier form) as one condition — Fitzpatrick's
-// "unary conditional expression" (Rule 6 / 7), mirroring Rust's
-// `rust_count_condition`. tree-sitter-ruby exposes the predicate via the
-// `condition` field for every form (block `if cond ... end` and modifier
-// `body if cond`), so the field lookup is position-independent. A bare
-// terminal (`if flag`) counts directly; a comparison / boolean chain
-// (`if a == b`, `if a && b`) is a nested `binary` node already counted by
-// the comparison-operator and `&&`/`||` walker arms, so it adds nothing
-// here. A parenthesised or negated predicate (`if (flag)`, `if !flag`) is
-// unwrapped by `ruby_inspect_container`. Without this arm, idiomatic Ruby
-// bare predicates reported 0 ABC conditions while Ruby's own cyclomatic
-// counted them, breaking the conditions >= decisions invariant
-// (#469/#473/#456); issue #696.
-fn ruby_count_condition(condition: &Node, conditions: &mut f64) {
-    use Ruby::*;
-    let kind = condition.kind_id().into();
-    if matches!(kind, ruby_bool_terminal_kinds!()) {
-        *conditions += 1.;
-    } else if matches!(
-        kind,
-        ParenthesizedStatements | Unary | Unary2 | Unary3 | Unary4 | Unary5
-    ) {
-        ruby_inspect_container(condition, conditions);
-    }
-}
-
-impl Abc for RubyCode {
-    fn compute<'a>(node: &Node<'a>, _code: &'a [u8], stats: &mut Stats) {
-        use Ruby::*;
-
-        match node.kind_id().into() {
-            Assignment | Assignment2 | OperatorAssignment | OperatorAssignment2 => {
-                stats.assignments += 1.;
-            }
-            // The bare predicate of every `if`/`unless`/`while`/`until`
-            // form (block and modifier) is one unary condition. The
-            // `condition` field locates the predicate position-
-            // independently across all eight node kinds (#696).
-            If | Unless | While | Until | IfModifier | UnlessModifier | WhileModifier
-            | UntilModifier => {
-                if let Some(cond) = node.child_by_field_name("condition") {
-                    ruby_count_condition(&cond, &mut stats.conditions);
-                }
-            }
-            Call | Call2 | Call3 | Call4 | Super | Yield | Yield2 => {
-                stats.branches += 1.;
-            }
-            EQEQ | BANGEQ | EQEQEQ | LT | GT | LTEQ | GTEQ | LTEQGT | EQTILDE | BANGTILDE
-            | Else | Elsif | When | QMARK | Rescue | RescueModifier | RescueModifier2
-            | RescueModifier3 => {
-                stats.conditions += 1.;
-            }
-            // Fitzpatrick Rule 9 walker: each non-comparison operand of a
-            // `&&` / `||` / `and` / `or` chain is one condition (issue
-            // #557). The short-circuit operators are not counted directly
-            // (cross-language policy, #395); the keyword forms `and` / `or`
-            // get the same treatment as `&&` / `||`.
-            AMPAMP | PIPEPIPE | And | Or => {
-                if let Some(parent) = node.parent() {
-                    ruby_count_unary_conditions(&parent, &mut stats.conditions);
-                }
-            }
-            _ => {}
-        }
-    }
-}
-
-// Fitzpatrick's ABC rules adapted for Python.
-//
-// - Assignments: every `Assignment` node that contains an explicit `=`
-//   token (plain assignment, walrus `:=` lives in `NamedExpression`,
-//   handled separately), plus every `AugmentedAssignment` (`+=`,
-//   `-=`, …) and every `NamedExpression` (walrus). Bare type-only
-//   annotations like `x: int` also parse as `Assignment` but have no
-//   `=` child — these are excluded so a class-level type annotation
-//   does not inflate the assignment count.
-// - Branches: every `Call` node. Python's "object construction" is
-//   syntactically a `Call` (`Foo()` parses as `call`), so the same
-//   arm covers it without a separate `New`-style case.
-// - Conditions: comparison operators (`ComparisonOperator` wraps
-//   `<`, `>`, `==`, `!=`, `is`, `is not`, `in`, `not in`, etc. as a
-//   single node), `ConditionalExpression` (ternary `a if c else b`),
-//   the unary `NotOperator` (paper's "unary conditional expression",
-//   Rule 7 / Figure 4), and the explicit arms of control flow:
-//   `ElifClause`, `ElseClause`, `ExceptClause`, `FinallyClause`,
-//   `CaseClause`. We do not separately count the `if` / `while`
-//   keyword: the condition expression itself is already covered by
-//   `ComparisonOperator`. This matches the token-level approach
-//   used for PHP / Bash.
-//
-//   `BooleanOperator` (Python's `and` / `or` wrapper) is
-//   deliberately NOT counted, and `NotOperator` is kept as the
-//   paper's "unary conditional expression". See the module-level
-//   `Stats` doc-comment for the cross-language `&&` / `||` policy
-//   (issue #395, walker tracked in #403).
-// Python ABC unary-conditional walker (Fitzpatrick Rule 9 / Figure 4;
-// issue #403). Python's `a and b` parses as `boolean_operator` (NOT
-// `binary_operator`), so the walker triggers on the `And` / `Or`
-// keyword tokens and iterates the immediate children of the parent
-// `boolean_operator`. Terminal-bool kinds: `Identifier`, `True`,
-// `False`, `Call`, `Attribute` (`obj.attr`), and `Subscript` (`xs[i]`).
-//
-// Unlike the C-family walkers, this one deliberately does NOT recurse
-// into `NotOperator`, `ComparisonOperator`, or nested `BooleanOperator`
-// children — each of those is counted by its own top-level dispatcher
-// arm. Re-counting them inside the walker would inflate the metric for
-// any `not x and y` / `x == 0 and y` shape. `ParenthesizedExpression`
-// is still unwrapped to catch bare-identifier operands like
-// `if (a) and b:` that the dispatcher would otherwise miss.
-fn python_inspect_container(container_node: &Node, conditions: &mut f64) {
-    use Python::*;
-
-    let mut node = *container_node;
-    let mut node_kind = node.kind_id().into();
-    let Some(parent) = node.parent() else { return };
-    let has_boolean_content = matches!(
-        parent.kind_id().into(),
-        BooleanOperator | IfStatement | WhileStatement | ConditionalExpression
-    );
-
-    loop {
-        if !matches!(node_kind, ParenthesizedExpression) {
-            break;
-        }
-
-        let Some(child) = node.child(1) else { break };
-        node = child;
-        node_kind = node.kind_id().into();
-
-        if matches!(node_kind, python_bool_terminal_kinds!()) {
-            if has_boolean_content {
-                *conditions += 1.;
-            }
-            break;
-        }
-    }
-}
-
-// Phase-2B (issue #403): Python `if` / `while` condition slot.
-// Python has no paren wrap around if-conditions and no top-level
-// terminal arm, so the condition has to be classified directly:
-//   - Identifier / True / False / Call / Attribute / Subscript at
-//     the top level counts once (Rule 6: bare-boolean condition).
-//   - ParenthesizedExpression unwraps via `python_inspect_container`.
-//   - NotOperator / ComparisonOperator / BooleanOperator are
-//     skipped: each is counted by its own top-level dispatcher arm
-//     (the `Or`/`And` keyword walker, the `NotOperator` arm, and
-//     the `ComparisonOperator` arm at lines `~1334-1390`).
-fn python_count_condition(condition: &Node, conditions: &mut f64) {
-    use Python::*;
-    let kind = condition.kind_id().into();
-    if matches!(kind, python_bool_terminal_kinds!()) {
-        *conditions += 1.;
-    } else if matches!(kind, ParenthesizedExpression) {
-        python_inspect_container(condition, conditions);
-    }
-}
-
-fn python_inspect_child(node: &Node, idx: usize, conditions: &mut f64) {
-    if let Some(child) = node.child(idx) {
-        python_count_condition(&child, conditions);
-    }
-}
-
-fn python_count_unary_conditions(list_node: &Node, conditions: &mut f64) {
-    use Python::*;
-
-    let list_kind = list_node.kind_id().into();
-    let mut cursor = list_node.cursor();
-
-    if cursor.goto_first_child() {
-        loop {
-            let node = cursor.node();
-            let node_kind = node.kind_id().into();
-
-            if matches!(node_kind, python_bool_terminal_kinds!())
-                && matches!(list_kind, BooleanOperator)
-            {
-                *conditions += 1.;
-            } else if matches!(node_kind, ParenthesizedExpression) {
-                python_inspect_container(&node, conditions);
-            }
-            // NotOperator / ComparisonOperator / nested BooleanOperator
-            // children are intentionally not walked here — each has
-            // its own top-level dispatcher arm and counting them
-            // again would double-count.
-
-            if !cursor.goto_next_sibling() {
-                break;
-            }
-        }
-    }
-}
-
-impl Abc for PythonCode {
-    fn compute<'a>(node: &Node<'a>, _code: &'a [u8], stats: &mut Stats) {
-        use Python::*;
-
-        match node.kind_id().into() {
-            // Plain `=` assignment. tree-sitter-python emits an
-            // `Assignment` node for both `x = 1` (LHS, `=`, RHS) and
-            // bare annotations `x: int` (LHS, `:`, type, *no* `=`).
-            // Filtering on the presence of an `EQ` child keeps the
-            // annotation-only case out of the count.
-            Assignment if node.first_child(|id| id == EQ).is_some() => {
-                stats.assignments += 1.;
-            }
-            // Augmented assignment (`+=`, `-=`, `*=`, …) always counts;
-            // walrus `name := expr` is a PEP-572 `NamedExpression` and
-            // also binds a value, so it counts as one assignment under
-            // Fitzpatrick's rule.
-            AugmentedAssignment | NamedExpression => {
-                stats.assignments += 1.;
-            }
-            // Every call — function call, method call, type
-            // construction — is one branch. Python parses `Foo()` as
-            // `Call`, so object construction folds into this arm.
-            Call => {
-                stats.branches += 1.;
-            }
-            // `x < y`, `a == b`, `c is None`, `n in xs`, `m not in xs`
-            // all parse as a single `ComparisonOperator` node — one
-            // node, one condition, regardless of how many comparison
-            // operators are chained.
-            ComparisonOperator
-            | ConditionalExpression
-            | ElifClause
-            | ElseClause
-            | ExceptClause
-            | FinallyClause
-            | NotOperator => {
-                // `NotOperator` is Python's unary `not`. Counting it
-                // mirrors Java's `!x` / C#'s `!x` Abc condition rule
-                // and closes the parity gap noted in #214 — without
-                // it, `if not flag:` reports 0 conditions while
-                // `if !flag` in Java reports 1. Nested combos like
-                // `not (x > 0)` count both the unary and the
-                // comparison once each (one logical "is-negation",
-                // one logical "comparison"), matching Java's
-                // `!(x > 0)`.
-                stats.conditions += 1.;
-            }
-            // A non-wildcard `case` arm contributes one condition,
-            // matching Rust's bare-`_` MatchArm filter and Java/C#'s
-            // `default:` rule. The bare wildcard is detected by: (a)
-            // `case_pattern` is `_`, AND (b) no `if_clause` sibling
-            // on the `case_clause` — `case _ if g:` carries a guard
-            // and still counts. The shared classifier lives in
-            // `super::npa` next to `pattern_is_bare_underscore`.
-            CaseClause if super::npa::python_case_clause_counts(node, UNDERSCORE as u16) => {
-                stats.conditions += 1.;
-            }
-            // Fitzpatrick Rule 9 walker: each operand of an `and` /
-            // `or` chain is one condition (issue #403). The `And` /
-            // `Or` keyword tokens live inside a `boolean_operator`
-            // wrapper which the walker iterates as the parent list.
-            And | Or => {
-                if let Some(parent) = node.parent() {
-                    python_count_unary_conditions(&parent, &mut stats.conditions);
-                }
-            }
-            // Phase-2B (issue #403): `if` / `while` condition slot.
-            // Python has no paren wrap around if-conditions, so the
-            // condition has to be classified directly. NotOperator,
-            // ComparisonOperator, BooleanOperator children each have
-            // their own dispatcher arms and are not re-counted here.
-            // `for` has no condition slot; ReturnStatement /
-            // ArgumentList do not need walker arms because every
-            // unary-conditional content node (NotOperator,
-            // ComparisonOperator) already has its own top-level arm.
-            IfStatement | WhileStatement => {
-                python_inspect_child(node, 1, &mut stats.conditions);
-            }
-            _ => {}
-        }
-    }
-}
-
-// Rust ABC unary-conditional walker (Fitzpatrick Rule 7 / Listing 2).
-//
-// On every `&&` / `||` token, we walk the parent `binary_expression`
-// and count each non-comparison operand as one condition. Identifier,
-// boolean literal, call, field-expression, and index-expression
-// operands count directly. Operands wrapped in `(…)` or `!…` route
-// through `rust_inspect_container`, which unwraps the wrapper chain
-// until it lands on a terminal — then counts. Operands that are
-// themselves nested `binary_expression`s (left-associative
-// `(a && b) && c`) are not counted at the outer site; the inner
-// `&&` token's own walker pass picks them up.
-//
-// The list-kind guard inside the count helper prevents an Identifier
-// or BooleanLiteral that happens to be an immediate child of a non-
-// binary parent from contributing — only direct operands of a
-// `binary_expression` count as unary conditions per Rule 7. See issue
-// #403.
-fn rust_inspect_container(container_node: &Node, conditions: &mut f64) {
-    use Rust::*;
-
-    let mut node = *container_node;
-    let mut node_kind = node.kind_id().into();
-    let Some(parent) = node.parent() else { return };
-    let mut has_boolean_content = matches!(
-        parent.kind_id().into(),
-        BinaryExpression | IfExpression | WhileExpression | LetChain | LetChain2
-    );
-
-    loop {
-        let is_parens = matches!(node_kind, ParenthesizedExpression);
-        let is_not = matches!(node_kind, UnaryExpression)
-            && node.child(0).is_some_and(|c| c.kind_id() == BANG as u16);
-
-        if !is_parens && !is_not {
-            break;
-        }
-        if !has_boolean_content && is_not {
-            has_boolean_content = true;
-        }
-
-        let Some(child) = node.child(1) else { break };
-        node = child;
-        node_kind = node.kind_id().into();
-
-        if matches!(node_kind, rust_bool_terminal_kinds!()) {
-            if has_boolean_content {
-                *conditions += 1.;
-            }
-            break;
-        }
-    }
-}
-
-// Phase-2B helpers (issue #403): classify a condition slot directly.
-// Used for the `if (cond)` / `while (cond)` / `return value` arms —
-// Fitzpatrick's Rule 6 / 7 ("unary conditional expression"). If the
-// condition itself is a terminal-bool kind (`if true {}`, `if a {}`),
-// it counts as one condition; if wrapped in `(...)` or `!...`,
-// `rust_inspect_container` unwraps until a terminal is found. Mirrors
-// the `java_count_condition` / `java_inspect_child` helper pair used
-// by `java_walk_ternary` / `java_walk_for_statement`.
-fn rust_count_condition(condition: &Node, conditions: &mut f64) {
-    use Rust::*;
-    let kind = condition.kind_id().into();
-    if matches!(kind, rust_bool_terminal_kinds!()) {
-        *conditions += 1.;
-    } else if matches!(kind, ParenthesizedExpression | UnaryExpression) {
-        rust_inspect_container(condition, conditions);
-    }
-}
-
-fn rust_inspect_child(node: &Node, idx: usize, conditions: &mut f64) {
-    if let Some(child) = node.child(idx) {
-        rust_count_condition(&child, conditions);
-    }
-}
-
-fn rust_count_unary_conditions(list_node: &Node, conditions: &mut f64) {
-    use Rust::*;
-
-    let list_kind = list_node.kind_id().into();
-    let mut cursor = list_node.cursor();
-
-    if cursor.goto_first_child() {
-        loop {
-            let node = cursor.node();
-            let node_kind = node.kind_id().into();
-
-            // Allow `LetChain` (and its hidden-rule alias `LetChain2`)
-            // alongside `BinaryExpression` as a known-boolean list
-            // parent: a Rust 2024 let-chain `if a && let Some(x) = b`
-            // makes `&&`'s parent the `LetChain` wrapper, not a
-            // `BinaryExpression`. Without this, bare-identifier
-            // operands inside a let-chain fall through and never
-            // contribute to the condition count, while their
-            // semantically equivalent `BinaryExpression` siblings do.
-            if matches!(node_kind, rust_bool_terminal_kinds!())
-                && matches!(list_kind, BinaryExpression | LetChain | LetChain2)
-            {
-                *conditions += 1.;
-            } else if node.is_named() {
-                rust_inspect_container(&node, conditions);
-            }
-
-            if !cursor.goto_next_sibling() {
-                break;
-            }
-        }
-    }
-}
-
-impl Abc for RustCode {
-    fn compute<'a>(node: &Node<'a>, _code: &'a [u8], stats: &mut Stats) {
-        use Rust::*;
-
-        match node.kind_id().into() {
-            // Plain `x = expr` (assignment_expression) and augmented
-            // forms `+=`, `-=`, `*=`, `/=`, `%=`, `&=`, `|=`, `^=`,
-            // `<<=`, `>>=` (compound_assignment_expr) both bind a
-            // value; each counts as one assignment. Rust grammar
-            // isolates both in distinct named nodes, so there is no
-            // risk of double-counting the contained `EQ` token here.
-            AssignmentExpression | CompoundAssignmentExpr => {
-                stats.assignments += 1.;
-            }
-            // `let x = expr;` and `let mut x = expr;` both carry an
-            // explicit `=` initializer — the `value` field is present
-            // on the `let_declaration` only when the initializer
-            // exists. Per Fitzpatrick (1997), every `=` operator
-            // increments A; the JS impl already counts `let x = 5;`
-            // (and excludes `const`). We follow the literal reading
-            // for Rust too and count both `let x = ...;` and
-            // `let mut x = ...;` — distinguishing the `mut` form
-            // would diverge from the JS rule (which does not
-            // distinguish `let` from `var`) and complicates the
-            // implementation without changing the cross-language
-            // story. Bare `let x;` (no initializer) leaves the
-            // `value` field unset and correctly stays out.
-            LetDeclaration if node.child_by_field_name("value").is_some() => {
-                stats.assignments += 1.;
-            }
-            // Every call expression — including method calls
-            // (`a.b.c()` parses as `call_expression` whose callee is a
-            // `field_expression`) — plus every `try_expression` (the
-            // `?` operator, a short-circuit return on Result / Option)
-            // contributes one branch. Macro invocations parse as
-            // `macro_invocation`, NOT `call_expression`, so they are
-            // intentionally NOT counted as branches.
-            CallExpression | TryExpression => {
-                stats.branches += 1.;
-            }
-            // Comparison operators emitted as token children of a
-            // `binary_expression`, `if let` / `while let` conditions,
-            // and the `else` keyword each count as one condition.
-            // `let_condition` covers both `if let` and `while let`
-            // (Rust's grammar uses the same node for both); inside a
-            // `let_chain` each `let_condition` counts separately.
-            // Java counts the `Else` token directly; Rust's grammar
-            // exposes the same token and we follow that lead.
-            LTEQ | GTEQ | EQEQ | BANGEQ | LetCondition | Else => {
-                stats.conditions += 1.;
-            }
-            // `<` / `>` doubles as type-argument delimiter; the
-            // `BinaryExpression` parent check disambiguates without
-            // needing to inspect siblings.
-            LT | GT
-                if node
-                    .parent()
-                    .is_some_and(|p| matches!(p.kind_id().into(), BinaryExpression)) =>
-            {
-                stats.conditions += 1.;
-            }
-            // Every non-wildcard `match_arm` is one condition. A bare
-            // `_ => ...` arm is the C / Java `default:` equivalent and
-            // is excluded — mirrors the cyclomatic treatment and
-            // Kotlin's `when` / Java's `case` rules. Patterns like
-            // `Some(_)`, `(_, x)`, or `_ if guard` are not bare
-            // wildcards and still count. The check scans only NAMED
-            // children of `match_pattern` so anonymous tokens like a
-            // leading `|` (allowed in or-patterns: `| _ => ...`) do
-            // not throw off the detection. A guard (`_ if g`) adds a
-            // second named child to `match_pattern` and so escapes
-            // the bare-wildcard filter.
-            MatchArm | MatchArm2 => {
-                let is_bare_wildcard = node.child_by_field_name("pattern").is_some_and(|pat| {
-                    super::npa::pattern_is_bare_underscore(&pat, UNDERSCORE as u16)
-                });
-                if !is_bare_wildcard {
-                    stats.conditions += 1.;
-                }
-            }
-            // Fitzpatrick Rule 7: each operand of a `&&` / `||` chain
-            // is one condition. The walker iterates immediate children
-            // of the parent `binary_expression`; the per-`&&` / per-`||`
-            // trigger keeps left-associative chains (`a && b && c`) at
-            // O(operands) total work since the inner operator's pass
-            // counts the inner pair and the outer operator's pass
-            // counts only the new outer operand. See issue #403.
-            AMPAMP | PIPEPIPE => {
-                if let Some(parent) = node.parent() {
-                    rust_count_unary_conditions(&parent, &mut stats.conditions);
-                }
-            }
-            // Phase-2B (issue #403): Fitzpatrick Rule 6 / 7 condition
-            // slots. `if true {}` / `if !a {}` count their condition
-            // once via `rust_count_condition` (terminal-at-top or
-            // paren / unary unwrap). The condition sits at child(1)
-            // for `if_expression` and `while_expression` (child(0)
-            // is the keyword). Rust has no `do_statement`, no
-            // ternary, and no for-condition slot.
-            IfExpression | WhileExpression => {
-                rust_inspect_child(node, 1, &mut stats.conditions);
-            }
-            // `return value;` — `value` sits at child(1). Use the
-            // bare `inspect_container` path (no top-level terminal
-            // count) so that `return x` reports zero conditions
-            // while `return !x` reports one. Matches Java's policy
-            // (`java_return_without_conditions`): a bare identifier
-            // in the return slot is not a unary conditional.
-            ReturnExpression => {
-                if let Some(value) = node.child(1) {
-                    rust_inspect_container(&value, &mut stats.conditions);
-                }
-            }
-            // Method-argument walker: `m(!a, !b)` contributes one
-            // condition per unary-conditional argument.
-            Arguments => {
-                rust_count_unary_conditions(node, &mut stats.conditions);
-            }
-            _ => {}
-        }
-    }
-}
-
-// Go ABC unary-conditional walker (issue #403; see `rust_inspect_container`
-// for the cross-language rationale). Terminal-bool kinds include calls,
-// selector access (`r.Field`), index access (`xs[i]`), and type
-// assertions (`x.(*T)`) — every kind whose evaluated value is implicitly
-// boolean in idiomatic Go for `if` / `for` conditions.
-fn go_inspect_container(container_node: &Node, conditions: &mut f64) {
-    use Go as G;
-
-    let mut node = *container_node;
-    let mut node_kind = node.kind_id().into();
-    let Some(parent) = node.parent() else { return };
-    let mut has_boolean_content = matches!(
-        parent.kind_id().into(),
-        G::BinaryExpression | G::IfStatement | G::ForStatement
-    );
-
-    loop {
-        let is_parens = matches!(node_kind, G::ParenthesizedExpression);
-        let is_not = matches!(node_kind, G::UnaryExpression)
-            && node.child(0).is_some_and(|c| c.kind_id() == G::BANG as u16);
-
-        if !is_parens && !is_not {
-            break;
-        }
-        if !has_boolean_content && is_not {
-            has_boolean_content = true;
-        }
-
-        let Some(child) = node.child(1) else { break };
-        node = child;
-        node_kind = node.kind_id().into();
-
-        if matches!(node_kind, go_bool_terminal_kinds!()) {
-            if has_boolean_content {
-                *conditions += 1.;
-            }
-            break;
-        }
-    }
-}
-
-// Phase-2B (issue #403): condition-slot dispatcher for Go.
-fn go_count_condition(condition: &Node, conditions: &mut f64) {
-    use Go as G;
-    let kind = condition.kind_id().into();
-    if matches!(kind, go_bool_terminal_kinds!()) {
-        *conditions += 1.;
-    } else if matches!(kind, G::ParenthesizedExpression | G::UnaryExpression) {
-        go_inspect_container(condition, conditions);
-    }
-}
-
-fn go_count_unary_conditions(list_node: &Node, conditions: &mut f64) {
-    use Go as G;
-
-    let list_kind = list_node.kind_id().into();
-    let mut cursor = list_node.cursor();
-
-    if cursor.goto_first_child() {
-        loop {
-            let node = cursor.node();
-            let node_kind = node.kind_id().into();
-
-            if matches!(node_kind, go_bool_terminal_kinds!())
-                && matches!(list_kind, G::BinaryExpression)
-            {
-                *conditions += 1.;
-            } else if node.is_named() {
-                go_inspect_container(&node, conditions);
-            }
-
-            if !cursor.goto_next_sibling() {
-                break;
-            }
-        }
-    }
-}
-
-impl Abc for GoCode {
-    fn compute<'a>(node: &Node<'a>, _code: &'a [u8], stats: &mut Stats) {
-        // Aliased because `Go::Go` (the `go` keyword variant) collides
-        // with the bare enum name in pattern position under
-        // `use Go::*;` (same workaround as in cyclomatic / cognitive).
-        use Go as G;
-
-        match node.kind_id().into() {
-            // Plain `=`, augmented `+=`, `-=`, … all parse as
-            // `assignment_statement`. `:=` is a short variable
-            // declaration. `x++` / `x--` rebind too.
-            G::AssignmentStatement | G::ShortVarDeclaration | G::IncStatement | G::DecStatement => {
-                stats.assignments += 1.;
-            }
-            // Every call expression — including method calls
-            // (`r.Method()` parses as `call_expression` whose callee is
-            // a `selector_expression`) — contributes one branch.
-            // Composite literals (`Point{X: 1}`) are NOT calls.
-            G::CallExpression => {
-                stats.branches += 1.;
-            }
-            // Comparison operators emitted as token children of a
-            // `binary_expression`, `else`, and each non-default switch
-            // / type-switch / select arm all contribute one condition.
-            // `<` / `>` double as type-argument delimiters in generic
-            // instantiations (`f[T any]`, `List[int]`); the
-            // `BinaryExpression` parent guard filters those out
-            // without inspecting siblings. `default_case` is
-            // intentionally excluded — like Java / C# `default:`, it
-            // does not introduce a new decision point.
-            G::EQEQ
-            | G::BANGEQ
-            | G::LTEQ
-            | G::GTEQ
-            | G::Else
-            | G::ExpressionCase
-            | G::TypeCase
-            | G::CommunicationCase => {
-                stats.conditions += 1.;
-            }
-            G::LT | G::GT
-                if node
-                    .parent()
-                    .is_some_and(|p| matches!(p.kind_id().into(), G::BinaryExpression)) =>
-            {
-                stats.conditions += 1.;
-            }
-            // Fitzpatrick Rule 7: each operand of a `&&` / `||` chain is
-            // one condition (issue #403). The walker iterates immediate
-            // children of the parent `binary_expression`.
-            G::AMPAMP | G::PIPEPIPE => {
-                if let Some(parent) = node.parent() {
-                    go_count_unary_conditions(&parent, &mut stats.conditions);
-                }
-            }
-            // Phase-2B (issue #403): Rule 6 / 7 condition slots.
-            // `if true {}` / `if !a {}` / `if (a) {}` count once.
-            // Use `child_by_field_name("condition")` so the
-            // `if x := f(); x { ... }` init-statement form is
-            // handled correctly — its `condition` field is at
-            // child(2) (not child(1), which is the init slot).
-            G::IfStatement => {
-                if let Some(cond) = node.child_by_field_name("condition") {
-                    go_count_condition(&cond, &mut stats.conditions);
-                }
-            }
-            // Phase-2B follow-up (findings.md #1): Go's `for` is its
-            // only loop with a bare-condition slot. Children:
-            //   `for cond {}`           → child(1) = condition
-            //   `for init; cond; post`  → child(1) = `for_clause`
-            //   `for range items {}`    → child(1) = `range_clause`
-            // `go_count_condition` filters non-terminal /
-            // non-paren / non-unary kinds, so `for_clause` and
-            // `range_clause` fall through harmlessly without a
-            // dedicated guard.
-            G::ForStatement => {
-                if let Some(cond) = node.child(1) {
-                    go_count_condition(&cond, &mut stats.conditions);
-                }
-            }
-            // `return value` — Go wraps the return values in an
-            // `expression_list` at child(1). Iterate the list's
-            // children and route each through `inspect_container`
-            // (NOT the terminal-at-top form): `return !x` counts
-            // the wrapped Identifier once, while `return x` (bare
-            // identifier in the return slot) reports zero
-            // conditions. Matches Java's policy in
-            // `java_return_without_conditions`. Bare `return`
-            // (no values) has no child(1).
-            G::ReturnStatement => {
-                if let Some(expr_list) = node.child(1) {
-                    for_each_named_child(&expr_list, &mut stats.conditions, go_inspect_container);
-                }
-            }
-            // Method-argument-list walker for `f(!a, !b)`. Two
-            // aliases — `argument_list` is emitted as ArgumentList
-            // or ArgumentList2 depending on production rule path.
-            G::ArgumentList | G::ArgumentList2 => {
-                go_count_unary_conditions(node, &mut stats.conditions);
-            }
-            _ => {}
-        }
-    }
-}
-
-// C++ ABC unary-conditional walker (Fitzpatrick Rule 9 in Figure 3;
-// see `rust_inspect_container` for the cross-language rationale).
-// Matches on node-kind NAMES so the helper is correct for every C-family
-// grammar: it is shared by the `CppCode` and `MozcppCode` ABC impls (#720),
-// and the Mozilla fork assigns different kind_ids to the same kinds (#732,
-// mirroring the npa fix in #731). Aliased kinds — `binary_expression`,
-// `parenthesized_expression`, `unary_expression` each have a second token-id
-// in the C++ grammar (under structured-binding / requires-clause production
-// rules) — all render to one base name, so a single string arm covers each
-// family: equivalent to the former `Cpp`-enum match for Cpp, and
-// grammar-agnostic for Mozcpp.
-fn cpp_inspect_container(container_node: &Node, conditions: &mut f64) {
-    let mut node = *container_node;
-    let mut node_kind = node.kind();
-    let Some(parent) = node.parent() else { return };
-    let parent_kind = parent.kind();
-    let mut has_boolean_content = matches!(
-        parent_kind,
-        "binary_expression" | "if_statement" | "while_statement" | "do_statement" | "for_statement"
-    ) || (parent_kind == "conditional_expression"
-        && node
-            .previous_sibling()
-            .is_none_or(|prev| !matches!(prev.kind(), "?" | ":")));
-
-    loop {
-        // `condition_clause` is the C++-grammar wrapper around an
-        // `if (...)` / `while (...)` head — same `(`, content, `)`
-        // shape as `parenthesized_expression`, so it unwraps the
-        // same way at child(1). `do { ... } while (...)`'s trailing
-        // condition is a plain `parenthesized_expression`.
-        let is_parens = matches!(node_kind, "parenthesized_expression" | "condition_clause");
-        let is_not =
-            node_kind == "unary_expression" && node.child(0).is_some_and(|c| c.kind() == "!");
-
-        if !is_parens && !is_not {
-            break;
-        }
-        if !has_boolean_content && is_not {
-            has_boolean_content = true;
-        }
-
-        let Some(child) = node.child(1) else { break };
-        node = child;
-        node_kind = node.kind();
-
-        if matches!(node_kind, cpp_bool_terminal_kinds!()) {
-            if has_boolean_content {
-                *conditions += 1.;
-            }
-            break;
-        }
-    }
-}
-
-// Phase-2B helpers (issue #403): condition-slot dispatcher for C++.
-// `if (cond)` / `while (cond)` / `return value` slots are
-// paren-wrapped in C++; `cpp_inspect_container` already handles the
-// `(...)` / `!...` unwrap chain and the boolean-context seed from
-// the parent kind. No top-level terminal counter is needed because
-// the paren wrapper provides the unwrap step.
-fn cpp_inspect_child(node: &Node, idx: usize, conditions: &mut f64) {
-    if let Some(child) = node.child(idx) {
-        cpp_inspect_container(&child, conditions);
-    }
-}
-
-fn cpp_count_unary_conditions(list_node: &Node, conditions: &mut f64) {
-    let list_kind = list_node.kind();
-    let mut cursor = list_node.cursor();
-
-    if cursor.goto_first_child() {
-        loop {
-            let node = cursor.node();
-            let node_kind = node.kind();
-
-            if matches!(node_kind, cpp_bool_terminal_kinds!()) && list_kind == "binary_expression" {
-                *conditions += 1.;
-            } else if node.is_named() {
-                cpp_inspect_container(&node, conditions);
-            }
-
-            if !cursor.goto_next_sibling() {
-                break;
-            }
-        }
-    }
-}
-
-impl Abc for CppCode {
-    fn compute<'a>(node: &Node<'a>, _code: &'a [u8], stats: &mut Stats) {
-        use Cpp::*;
-
-        match node.kind_id().into() {
-            // `assignment_expression` covers both plain `=` and every
-            // compound form (`+=`, `-=`, `*=`, `/=`, `%=`, `&=`, `|=`,
-            // `^=`, `<<=`, `>>=`); the grammar lifts them all into a
-            // single named node so we count once per
-            // `assignment_expression`. `update_expression` covers both
-            // prefix and postfix `++` / `--`.
-            AssignmentExpression | AssignmentExpression2 | UpdateExpression => {
-                stats.assignments += 1.;
-            }
-            // `int x = expr;` parses as a `declaration` carrying an
-            // `init_declarator` of the form `declarator = value`. Per
-            // Fitzpatrick (1997), every `=` operator increments A; the
-            // JS impl already counts `let x = 5;` (and excludes
-            // `const`). We follow the literal reading for C++ too and
-            // count every `init_declarator` whose body contains an
-            // explicit `=` token. `const int x = 5;` is counted along
-            // with `int x = 5;` — distinguishing them would diverge
-            // from the JS rule's "let counted, const not" mapping
-            // because C++ `const` semantics are unlike JS `const` (a
-            // C++ `const int x` binding is the canonical "one
-            // assignment to initialise" — closer to Rust's
-            // non-`mut` `let` than to JS's hoisted reference binding).
-            // `int x;` parses as a plain declarator inside the
-            // `declaration`, not an `init_declarator`, so this arm
-            // never fires for un-initialised declarations. The second
-            // `init_declarator` grammar form `int x(5);` / `int x{5};`
-            // (paren / brace init) carries no `=` token and stays out
-            // — only the `=` operator counts.
-            InitDeclarator if node.first_child(|id| id == EQ as u16).is_some() => {
-                stats.assignments += 1.;
-            }
-            // Every call counts (method calls fold in as
-            // `call_expression` with a `field_expression` callee). The
-            // C++ grammar exposes two aliased `call_expression` ids.
-            // `new T(...)` allocations count as a branch — they invoke
-            // a constructor, mirroring Java's `New` and C#'s
-            // `ObjectCreationExpression` rule.
-            CallExpression | CallExpression2 | NewExpression => {
-                stats.branches += 1.;
-            }
-            // Comparison operators emitted as token children of a
-            // `binary_expression`. The C++20 spaceship `<=>` (`LTEQGT`)
-            // is a comparison operator and counts once per use.
-            // `else` opens an alternative branch path; `case`
-            // (non-default) adds one per switch arm; `?` opens a
-            // ternary; `try` / `catch` count per Fitzpatrick (and
-            // Java's rule). `Try2` is the second token-id alias the
-            // C++ grammar emits for `try` (it appears under
-            // structured-exception forms).
-            //
-            // `&&` / `||` are deliberately NOT counted (Fitzpatrick
-            // Rule 7 in Figure 3 for C++; the unary-conditional
-            // counterpart is Rule 9). See the module-level `Stats`
-            // doc-comment for the cross-language policy (issue
-            // #395, walker tracked in #403).
-            LTEQ | GTEQ | EQEQ | BANGEQ | LTEQGT | Else | Case | QMARK | Try | Try2 | Catch => {
-                stats.conditions += 1.;
-            }
-            // Plain `<` / `>` doubles as template-argument and
-            // template-parameter delimiter (`std::vector<int>`,
-            // `template <typename T>`). The `binary_expression` parent
-            // check disambiguates without inspecting siblings — only
-            // comparison uses of `<` / `>` count. Both kind-id aliases
-            // (`BinaryExpression`, `BinaryExpression2`) are accepted
-            // because the C++ grammar emits the same node under two
-            // production-rule paths.
-            LT | GT
-                if node.parent().is_some_and(|p| {
-                    matches!(p.kind_id().into(), BinaryExpression | BinaryExpression2)
-                }) =>
-            {
-                stats.conditions += 1.;
-            }
-            // Fitzpatrick Rule 9 (C++ in Figure 3): each operand of a
-            // `&&` / `||` chain is one condition (issue #403).
-            AMPAMP | PIPEPIPE => {
-                if let Some(parent) = node.parent() {
-                    cpp_count_unary_conditions(&parent, &mut stats.conditions);
-                }
-            }
-            // Phase-2B (issue #403): condition slots. C++ wraps every
-            // `if (...)` / `while (...)` / `do {…} while (...)` /
-            // `return value` in a paren / parenthesized expression
-            // (return is unparenthesized but its child(1) is the
-            // expression). `cpp_inspect_container` handles the
-            // `(...)` / `!...` unwrap so `if (true)` and `return !x`
-            // each count one condition; bare `return x` reports zero.
-            // Use `child_by_field_name("condition")` for if/while so
-            // the `if constexpr (cond)` form (where child(1) is the
-            // `constexpr` keyword, not the condition_clause) is
-            // handled correctly. Return uses positional child(1)
-            // — its value field is always at index 1, no optional
-            // attribute precedes it.
-            IfStatement | WhileStatement => {
-                if let Some(cond) = node.child_by_field_name("condition") {
-                    cpp_inspect_container(&cond, &mut stats.conditions);
-                }
-            }
-            ReturnStatement => {
-                cpp_inspect_child(node, 1, &mut stats.conditions);
-            }
-            // `do { ... } while (cond);` — children: `do`, body,
-            // `while`, condition (parenthesized). Condition at child(3).
-            DoStatement => {
-                cpp_inspect_child(node, 3, &mut stats.conditions);
-            }
-            // `f(!a, !b)` — argument list walker. Two aliases —
-            // `argument_list` is emitted as ArgumentList or
-            // ArgumentList2 depending on production rule path.
-            ArgumentList | ArgumentList2 => {
-                cpp_count_unary_conditions(node, &mut stats.conditions);
-            }
-            _ => {}
-        }
-    }
-}
-
-impl Abc for CCode {
-    fn compute<'a>(node: &Node<'a>, _code: &'a [u8], stats: &mut Stats) {
-        use C::*;
-
-        match node.kind_id().into() {
-            // `assignment_expression` covers both plain `=` and every
-            // compound form (`+=`, `-=`, `*=`, `/=`, `%=`, `&=`, `|=`,
-            // `^=`, `<<=`, `>>=`); the grammar lifts them all into a
-            // single named node so we count once per
-            // `assignment_expression`. `update_expression` covers both
-            // prefix and postfix `++` / `--`.
-            AssignmentExpression | UpdateExpression => {
-                stats.assignments += 1.;
-            }
-            // `int x = expr;` parses as a `declaration` carrying an
-            // `init_declarator` of the form `declarator = value`. Per
-            // Fitzpatrick (1997), every `=` operator increments A, so we
-            // count every `init_declarator` whose body contains an
-            // explicit `=` token (`const int x = 5;` counts like
-            // `int x = 5;`). `int x;` parses as a plain declarator inside
-            // the `declaration`, not an `init_declarator`, so this arm
-            // never fires for un-initialised declarations.
-            InitDeclarator if node.first_child(|id| id == EQ as u16).is_some() => {
-                stats.assignments += 1.;
-            }
-            // Every call counts. The C grammar exposes two aliased
-            // `call_expression` ids. C has no `new` allocations, so
-            // (unlike the C++ impl) there is no `NewExpression` branch.
-            CallExpression | CallExpression2 => {
-                stats.branches += 1.;
-            }
-            // Comparison operators emitted as token children of a
-            // `binary_expression`. `else` opens an alternative branch
-            // path; `case` (non-default) adds one per switch arm; `?`
-            // opens a ternary. C has no exceptions and no `<=>`
-            // spaceship, so — unlike the C++ impl — there are no
-            // `try` / `catch` / `LTEQGT` condition arms.
-            //
-            // `&&` / `||` are deliberately NOT counted (Fitzpatrick
-            // Rule 7 in Figure 3; the unary-conditional
-            // counterpart is Rule 9). See the module-level `Stats`
-            // doc-comment for the cross-language policy (issue
-            // #395, walker tracked in #403).
-            LTEQ | GTEQ | EQEQ | BANGEQ | Else | Case | QMARK => {
-                stats.conditions += 1.;
-            }
-            // Plain `<` / `>` are comparison operators (C has no
-            // templates, so there is no template-delimiter ambiguity to
-            // resolve). The `binary_expression` parent check ensures only
-            // comparison uses count. Both kind-id aliases
-            // (`BinaryExpression`, `BinaryExpression2`) are accepted
-            // because the grammar emits the node under two
-            // production-rule paths.
-            LT | GT
-                if node.parent().is_some_and(|p| {
-                    matches!(p.kind_id().into(), BinaryExpression | BinaryExpression2)
-                }) =>
-            {
-                stats.conditions += 1.;
-            }
-            // Fitzpatrick Rule 9 (Figure 3): each operand of a
-            // `&&` / `||` chain is one condition (issue #403).
-            AMPAMP | PIPEPIPE => {
-                if let Some(parent) = node.parent() {
-                    cpp_count_unary_conditions(&parent, &mut stats.conditions);
-                }
-            }
-            // Phase-2B (issue #403): condition slots. C wraps every
-            // `if (...)` / `while (...)` / `do {…} while (...)` /
-            // `return value` in a paren / parenthesized expression
-            // (return is unparenthesized but its child(1) is the
-            // expression). `cpp_inspect_container` handles the
-            // `(...)` / `!...` unwrap so `if (true)` and `return !x`
-            // each count one condition; bare `return x` reports zero.
-            // Use `child_by_field_name("condition")` for if/while (C has
-            // no `if constexpr`, so child(1) is always the
-            // condition_clause). Return uses positional child(1) — its
-            // value field is always at index 1.
-            IfStatement | WhileStatement => {
-                if let Some(cond) = node.child_by_field_name("condition") {
-                    cpp_inspect_container(&cond, &mut stats.conditions);
-                }
-            }
-            ReturnStatement => {
-                cpp_inspect_child(node, 1, &mut stats.conditions);
-            }
-            // `do { ... } while (cond);` — children: `do`, body,
-            // `while`, condition (parenthesized). Condition at child(3).
-            DoStatement => {
-                cpp_inspect_child(node, 3, &mut stats.conditions);
-            }
-            // `f(!a, !b)` — argument list walker. Two aliases —
-            // `argument_list` is emitted as ArgumentList or
-            // ArgumentList2 depending on production rule path.
-            ArgumentList | ArgumentList2 => {
-                cpp_count_unary_conditions(node, &mut stats.conditions);
-            }
-            _ => {}
-        }
-    }
-}
-
-// Objective-C ABC. ObjC is C plus message sends and `@`-directives, so
-// the walker is the C/C++ shape (it reuses the grammar-agnostic
-// `cpp_inspect_container` / `cpp_inspect_child` / `cpp_count_unary_conditions`
-// helpers, which match on node-kind strings) with two ObjC additions:
-//   * the `B` (branch) dimension counts `message_expression`
-//     (`[obj msg:x]`) alongside C `call_expression`s — a message send is
-//     a call. ObjC has no `new` allocator (`[Foo alloc]` is itself a
-//     message send), so there is no `NewExpression` arm.
-//   * the `C` (condition) dimension counts `@try` (the `@try` keyword)
-//     and each `@catch` handler (`catch_clause`), mirroring how the C++
-//     impl counts `try` / `catch`. ObjC has no bare `catch` keyword token
-//     (it is `@catch` + a `catch_clause` node) and no `<=>` spaceship.
-//     `@throw` is not a condition (the C++ impl likewise does not count
-//     `throw`).
-impl Abc for ObjcCode {
-    fn compute<'a>(node: &Node<'a>, _code: &'a [u8], stats: &mut Stats) {
-        use Objc::*;
-
-        match node.kind_id().into() {
-            AssignmentExpression | UpdateExpression => {
-                stats.assignments += 1.;
-            }
-            InitDeclarator if node.first_child(|id| id == EQ as u16).is_some() => {
-                stats.assignments += 1.;
-            }
-            // Every call counts: C `call_expression` (two aliased ids).
-            CallExpression | CallExpression2 => {
-                stats.branches += 1.;
-            }
-            // An ObjC message send `[obj msg:x]` is also a call (B). Its
-            // arguments are direct children of `message_expression` (there
-            // is no `argument_list` wrapper, unlike a C call), so unary
-            // conditions in them (`[obj msg:!x]`, Fitzpatrick Rule 9) are
-            // inspected here rather than via the `ArgumentList` arm below.
-            MessageExpression => {
-                stats.branches += 1.;
-                cpp_count_unary_conditions(node, &mut stats.conditions);
-            }
-            // Comparison operators, `else` / `case` / `?` branch openers,
-            // and the `@try` / `@catch` exception conditions. `&&` / `||`
-            // are deliberately excluded (Fitzpatrick Rule 7; the
-            // unary-conditional counterpart is Rule 9, handled below).
-            LTEQ | GTEQ | EQEQ | BANGEQ | Else | Case | QMARK | ATtry | CatchClause => {
-                stats.conditions += 1.;
-            }
-            // Plain `<` / `>` count only in comparison position (parent is
-            // a `binary_expression`). ObjC has no templates, but the parent
-            // check is kept for parity with the C/C++ impl.
-            LT | GT
-                if node.parent().is_some_and(|p| {
-                    matches!(p.kind_id().into(), BinaryExpression | BinaryExpression2)
-                }) =>
-            {
-                stats.conditions += 1.;
-            }
-            AMPAMP | PIPEPIPE => {
-                if let Some(parent) = node.parent() {
-                    cpp_count_unary_conditions(&parent, &mut stats.conditions);
-                }
-            }
-            IfStatement | WhileStatement => {
-                if let Some(cond) = node.child_by_field_name("condition") {
-                    cpp_inspect_container(&cond, &mut stats.conditions);
-                }
-            }
-            ReturnStatement => {
-                cpp_inspect_child(node, 1, &mut stats.conditions);
-            }
-            DoStatement => {
-                cpp_inspect_child(node, 3, &mut stats.conditions);
-            }
-            ArgumentList | ArgumentList2 => {
-                cpp_count_unary_conditions(node, &mut stats.conditions);
-            }
-            _ => {}
-        }
-    }
-}
-
-impl Abc for MozcppCode {
-    fn compute<'a>(node: &Node<'a>, _code: &'a [u8], stats: &mut Stats) {
-        use Mozcpp::*;
-
-        match node.kind_id().into() {
-            // `assignment_expression` covers both plain `=` and every
-            // compound form (`+=`, `-=`, `*=`, `/=`, `%=`, `&=`, `|=`,
-            // `^=`, `<<=`, `>>=`); the grammar lifts them all into a
-            // single named node so we count once per
-            // `assignment_expression`. `update_expression` covers both
-            // prefix and postfix `++` / `--`.
-            AssignmentExpression | AssignmentExpression2 | UpdateExpression => {
-                stats.assignments += 1.;
-            }
-            // `int x = expr;` parses as a `declaration` carrying an
-            // `init_declarator` of the form `declarator = value`. Per
-            // Fitzpatrick (1997), every `=` operator increments A; the
-            // JS impl already counts `let x = 5;` (and excludes
-            // `const`). We follow the literal reading for C++ too and
-            // count every `init_declarator` whose body contains an
-            // explicit `=` token. `const int x = 5;` is counted along
-            // with `int x = 5;` — distinguishing them would diverge
-            // from the JS rule's "let counted, const not" mapping
-            // because C++ `const` semantics are unlike JS `const` (a
-            // C++ `const int x` binding is the canonical "one
-            // assignment to initialise" — closer to Rust's
-            // non-`mut` `let` than to JS's hoisted reference binding).
-            // `int x;` parses as a plain declarator inside the
-            // `declaration`, not an `init_declarator`, so this arm
-            // never fires for un-initialised declarations. The second
-            // `init_declarator` grammar form `int x(5);` / `int x{5};`
-            // (paren / brace init) carries no `=` token and stays out
-            // — only the `=` operator counts.
-            InitDeclarator if node.first_child(|id| id == EQ as u16).is_some() => {
-                stats.assignments += 1.;
-            }
-            // Every call counts (method calls fold in as
-            // `call_expression` with a `field_expression` callee). The
-            // C++ grammar exposes two aliased `call_expression` ids.
-            // `new T(...)` allocations count as a branch — they invoke
-            // a constructor, mirroring Java's `New` and C#'s
-            // `ObjectCreationExpression` rule.
-            CallExpression | CallExpression2 | NewExpression => {
-                stats.branches += 1.;
-            }
-            // Comparison operators emitted as token children of a
-            // `binary_expression`. The C++20 spaceship `<=>` (`LTEQGT`)
-            // is a comparison operator and counts once per use.
-            // `else` opens an alternative branch path; `case`
-            // (non-default) adds one per switch arm; `?` opens a
-            // ternary; `try` / `catch` count per Fitzpatrick (and
-            // Java's rule). `Try2` is the second token-id alias the
-            // C++ grammar emits for `try` (it appears under
-            // structured-exception forms).
-            //
-            // `&&` / `||` are deliberately NOT counted (Fitzpatrick
-            // Rule 7 in Figure 3 for C++; the unary-conditional
-            // counterpart is Rule 9). See the module-level `Stats`
-            // doc-comment for the cross-language policy (issue
-            // #395, walker tracked in #403).
-            LTEQ | GTEQ | EQEQ | BANGEQ | LTEQGT | Else | Case | QMARK | Try | Try2 | Catch => {
-                stats.conditions += 1.;
-            }
-            // Plain `<` / `>` doubles as template-argument and
-            // template-parameter delimiter (`std::vector<int>`,
-            // `template <typename T>`). The `binary_expression` parent
-            // check disambiguates without inspecting siblings — only
-            // comparison uses of `<` / `>` count. Both kind-id aliases
-            // (`BinaryExpression`, `BinaryExpression2`) are accepted
-            // because the C++ grammar emits the same node under two
-            // production-rule paths.
-            LT | GT
-                if node.parent().is_some_and(|p| {
-                    matches!(p.kind_id().into(), BinaryExpression | BinaryExpression2)
-                }) =>
-            {
-                stats.conditions += 1.;
-            }
-            // Fitzpatrick Rule 9 (C++ in Figure 3): each operand of a
-            // `&&` / `||` chain is one condition (issue #403).
-            AMPAMP | PIPEPIPE => {
-                if let Some(parent) = node.parent() {
-                    cpp_count_unary_conditions(&parent, &mut stats.conditions);
-                }
-            }
-            // Phase-2B (issue #403): condition slots. C++ wraps every
-            // `if (...)` / `while (...)` / `do {…} while (...)` /
-            // `return value` in a paren / parenthesized expression
-            // (return is unparenthesized but its child(1) is the
-            // expression). `cpp_inspect_container` handles the
-            // `(...)` / `!...` unwrap so `if (true)` and `return !x`
-            // each count one condition; bare `return x` reports zero.
-            // Use `child_by_field_name("condition")` for if/while so
-            // the `if constexpr (cond)` form (where child(1) is the
-            // `constexpr` keyword, not the condition_clause) is
-            // handled correctly. Return uses positional child(1)
-            // — its value field is always at index 1, no optional
-            // attribute precedes it.
-            IfStatement | WhileStatement => {
-                if let Some(cond) = node.child_by_field_name("condition") {
-                    cpp_inspect_container(&cond, &mut stats.conditions);
-                }
-            }
-            ReturnStatement => {
-                cpp_inspect_child(node, 1, &mut stats.conditions);
-            }
-            // `do { ... } while (cond);` — children: `do`, body,
-            // `while`, condition (parenthesized). Condition at child(3).
-            DoStatement => {
-                cpp_inspect_child(node, 3, &mut stats.conditions);
-            }
-            // `f(!a, !b)` — argument list walker. Two aliases —
-            // `argument_list` is emitted as ArgumentList or
-            // ArgumentList2 depending on production rule path.
-            ArgumentList | ArgumentList2 => {
-                cpp_count_unary_conditions(node, &mut stats.conditions);
-            }
-            _ => {}
-        }
-    }
-}
-
-impl Abc for BashCode {
-    fn compute<'a>(node: &Node<'a>, code: &'a [u8], stats: &mut Stats) {
-        match node.kind_id().into() {
-            // Each `variable_assignment` is one assignment regardless of
-            // operator (`=`, `+=`, `-=`, …) — counting the parent node
-            // avoids double-counting `Bash::EQ`, which is also produced
-            // for the `=` inside `[ a = b ]` test expressions.
-            Bash::VariableAssignment | Bash::VariableAssignment2 => {
-                stats.assignments += 1.;
-            }
-            // Every command invocation is a branch in the ABC sense
-            // (function-call / message-pass). `return` and `exit` builtins
-            // are also `Bash::Command` nodes and count here too.
-            Bash::Command => {
-                stats.branches += 1.;
-            }
-            // Two condition signals share this arm:
-            //
-            // - Comparison operators inside `[[ … ]]` and `(( … ))`, plus
-            //   the prefix test operators `-z`, `-n`, `-eq`, `-lt`, … which
-            //   the grammar emits as `Bash::TestOperator`.
-            // - Control-flow branches (`if`/`elif`/`while`). A Bash predicate
-            //   is a command, so the branch keyword itself is the only
-            //   condition signal. These branch keywords mirror the matching
-            //   Bash cyclomatic decisions (`if`/`elif`/`while`; not `for`/
-            //   `&&`/`||`), lifting ABC off 0 for `if cmd; then … elif … fi`.
-            Bash::EQEQ
-            | Bash::BANGEQ
-            | Bash::LT
-            | Bash::GT
-            | Bash::LTEQ
-            | Bash::GTEQ
-            | Bash::EQTILDE
-            | Bash::TestOperator
-            | Bash::IfStatement
-            | Bash::ElifClause
-            | Bash::WhileStatement => {
-                stats.conditions += 1.;
-            }
-            // Case arms are conditions too, but counted per-arm like
-            // cyclomatic, excluding the bare-`*)` wildcard (the Bash analogue
-            // of `default:`) (#696).
-            Bash::CaseItem | Bash::CaseItem2
-                if !crate::metrics::cyclomatic::bash_case_item_is_bare_wildcard(node, code) =>
-            {
-                stats.conditions += 1.;
-            }
-            _ => {}
-        }
-    }
-}
-
-// ABC token-level helpers for Java. Each helper covers one of the four
-// categories ABC tracks (assignments / branches / conditions / walked
-// unary conditions). Each returns `true` when it owns the node so the
-// dispatcher in `impl Abc for JavaCode::compute` can short-circuit and
-// avoid re-matching the same kind across categories. The arms are
-// mutually exclusive in the source language so a short-circuit chain
-// reproduces the original `match` semantics bit-for-bit.
-
-// Shared helper: passes `node.child(idx)` to `java_inspect_container`.
-// The container helper is a no-op on kinds other than
-// `ParenthesizedExpression` / `!`-prefixed `UnaryExpression`, so no
-// `matches!` guard is needed at the call site.
-fn java_inspect_child(node: &Node, idx: usize, conditions: &mut f64) {
-    if let Some(child) = node.child(idx) {
-        java_inspect_container(&child, conditions);
-    }
-}
-
-// Counts assignment tokens and maintains the declaration-kind stack
-// used to suppress `=` inside `final`-marked declarations.
-fn java_count_token_assignment(node: &Node, stats: &mut Stats) -> bool {
-    use Java::*;
-    match node.kind_id().into() {
-        STAREQ | SLASHEQ | PERCENTEQ | DASHEQ | PLUSEQ | LTLTEQ | GTGTEQ | AMPEQ | PIPEEQ
-        | CARETEQ | GTGTGTEQ | PLUSPLUS | DASHDASH => {
-            stats.assignments += 1.;
-        }
-        FieldDeclaration | LocalVariableDeclaration => {
-            stats.declaration.push(DeclKind::Var);
-        }
-        Final => {
-            if let Some(DeclKind::Var) = stats.declaration.last() {
-                stats.declaration.push(DeclKind::Const);
-            }
-        }
-        SEMI => {
-            if let Some(DeclKind::Const | DeclKind::Var) = stats.declaration.last() {
-                stats.declaration.clear();
-            }
-        }
-        // Excludes constant declarations (top of stack == Const).
-        EQ => {
-            if stats
-                .declaration
-                .last()
-                .is_none_or(|decl| matches!(decl, DeclKind::Var))
-            {
-                stats.assignments += 1.;
-            }
-        }
-        _ => return false,
-    }
-    true
-}
-
-// Counts branch tokens: every method call or `new` allocation.
-fn java_count_token_branch(node: &Node, stats: &mut Stats) -> bool {
-    use Java::*;
-    if matches!(node.kind_id().into(), MethodInvocation | New) {
-        stats.branches += 1.;
-        return true;
-    }
-    false
-}
-
-// Counts condition tokens: comparison operators, control-flow keywords,
-// and `<` / `>` outside generic-type contexts. The `default` arm of a
-// `switch` is excluded: it is the unconditional fallthrough, so
-// cyclomatic counts only the `Case` arms (issue #469). Java's classic
-// statement switch (`default:`) and arrow switch (`default ->`) both
-// emit the same `Default` token under `switch_label`, so omitting it
-// here covers both forms.
-fn java_count_token_condition(node: &Node, stats: &mut Stats) -> bool {
-    use Java::*;
-    match node.kind_id().into() {
-        GTEQ | LTEQ | EQEQ | BANGEQ | Else | Case | QMARK | Try | Catch => {
-            stats.conditions += 1.;
-        }
-        // Excludes `<` / `>` used for generic types (`Box<T>`).
-        GT | LT => {
-            if let Some(parent) = node.parent()
-                && !matches!(parent.kind_id().into(), TypeArguments)
-            {
-                stats.conditions += 1.;
-            }
-        }
-        _ => return false,
-    }
-    true
-}
-
-fn java_walk_for_conditions(node: &Node, stats: &mut Stats) {
-    use Java::*;
-    let conds = &mut stats.conditions;
-    match node.kind_id().into() {
-        // Unary conditions in elements separated by `&&` / `||`.
-        AMPAMP | PIPEPIPE => {
-            if let Some(parent) = node.parent() {
-                java_count_unary_conditions(&parent, conds);
-            }
-        }
-        // Unary conditions among method arguments.
-        ArgumentList => java_count_unary_conditions(node, conds),
-        // Child 1: `if (cond) ...`, `while (cond) ...`, `return value;`.
-        IfStatement | WhileStatement | ReturnStatement => java_inspect_child(node, 1, conds),
-        // Child 2: assignment / declarator RHS, lambda body
-        // (`params -> body`).
-        VariableDeclarator | AssignmentExpression | LambdaExpression => {
-            java_inspect_child(node, 2, conds);
-        }
-        // Child 3: the `while (cond)` condition of `do { ... } while (...);`.
-        DoStatement => java_inspect_child(node, 3, conds),
-        TernaryExpression => java_walk_ternary(node, stats),
-        ForStatement => java_walk_for_statement(node, stats),
-        _ => {}
-    }
-}
-
-fn java_walk_ternary(node: &Node, stats: &mut Stats) {
-    use Java::*;
-    let conds = &mut stats.conditions;
-    // Child 0: condition itself. The terminal set mirrors the one in
-    // `java_inspect_container` (issue #372 / lesson #19).
-    if let Some(condition) = node.child(0) {
-        match condition.kind_id().into() {
-            java_bool_terminal_kinds!() => *conds += 1.,
-            ParenthesizedExpression | UnaryExpression => {
-                java_inspect_container(&condition, conds);
-            }
-            _ => {}
-        }
-    }
-    // Children 2 and 4: the two branch expressions.
-    java_inspect_child(node, 2, conds);
-    java_inspect_child(node, 4, conds);
-}
-
-// Handles the `for (...)` multi-shape positional cascade: the loop
-// condition lives at child 3 when the initializer is a variable
-// declaration, otherwise at child 4 (split off by the `;` at child 3).
-fn java_walk_for_statement(node: &Node, stats: &mut Stats) {
-    use Java::*;
-    let Some(condition) = node.child(3) else {
-        return;
-    };
-    // Terminal set mirrors `java_inspect_container` (issue #372 / lesson
-    // #19): FieldAccess / CastExpression / ArrayAccess /
-    // InstanceofExpression all evaluate to a boolean in idiomatic Java
-    // for-condition slots.
-    match condition.kind_id().into() {
-        // `for (i = 0; cond; ...)` — initializer is an expression, so
-        // the leading `;` sits at child 3 and the real condition at 4.
-        SEMI => {
-            if let Some(cond) = node.child(4) {
-                match cond.kind_id().into() {
-                    // The terminal set here is `java_bool_terminal_kinds!()`
-                    // augmented with `SEMI` / `RPAREN` to handle the
-                    // empty-condition `for (init; ; ...)` and `for (;;)`
-                    // forms: the trailing `;` or closing `)` lands at
-                    // child(4) and is treated as a vacuously-true
-                    // condition. The arm cannot be written as
-                    // `java_bool_terminal_kinds!() | SEMI | RPAREN`
-                    // because clippy's `unnested_or_patterns` forbids
-                    // mixing a macro-expanded OR with extra literal
-                    // patterns, and splitting into two arms with the
-                    // same body trips `match_same_arms`.
-                    MethodInvocation | Identifier | True | False | FieldAccess | CastExpression
-                    | ArrayAccess | InstanceofExpression | SEMI | RPAREN => {
-                        stats.conditions += 1.;
-                    }
-                    ParenthesizedExpression | UnaryExpression => {
-                        java_inspect_container(&cond, &mut stats.conditions);
-                    }
-                    _ => {}
-                }
-            }
-        }
-        java_bool_terminal_kinds!() => {
-            stats.conditions += 1.;
-        }
-        ParenthesizedExpression | UnaryExpression => {
-            java_inspect_container(&condition, &mut stats.conditions);
-        }
-        _ => {}
-    }
-}
-
-// Fitzpatrick, Jerry (1997). "Applying the ABC metric to C, C++ and Java". C++ Report.
-// Source: https://www.softwarerenovation.com/Articles.aspx
-// ABC Java rules: (page 8, figure 4)
-// ABC Java example: (page 15, listing 4)
-impl Abc for JavaCode {
-    // Short-circuit chain across four mutually-exclusive category
-    // helpers. Each helper returns `true` when it owns the node, so
-    // the dispatcher early-exits to avoid re-matching the same kind in
-    // a later helper. The original pre-refactor `match` enforced
-    // one-arm-per-kind by construction; this chain preserves the same
-    // semantics only as long as no node kind is matched by more than
-    // one helper. If you add a new arm covering a kind already matched
-    // by an earlier helper, the earlier helper's `return` will silently
-    // hide it — split the kinds across helpers explicitly instead.
-    fn compute<'a>(node: &Node<'a>, _code: &'a [u8], stats: &mut Stats) {
-        if java_count_token_assignment(node, stats) {
-            return;
-        }
-        if java_count_token_branch(node, stats) {
-            return;
-        }
-        if java_count_token_condition(node, stats) {
-            return;
-        }
-        java_walk_for_conditions(node, stats);
-    }
-}
-
-// ABC token-level helpers for Groovy. Mirrors the Java helper layout
-// (assignments / branches / conditions / walked) with the dekobon
-// Groovy grammar's specific deltas — `CommandChain` as a branch
-// alongside `MethodInvocation` (#247); `DoWhileStatement` replacing
-// Java's `DoStatement`; no `LambdaExpression` (Groovy closures take
-// block bodies, no implicit-return arm); and `if (…)` / `while (…)` /
-// `do { … } while (…)` parens inlined as token children rather than
-// wrapped in `parenthesized_expression`, so the condition sits at a
-// different child index and goes through `groovy_count_condition`.
-
-// Groovy mirror of `java_inspect_child`: passes `node.child(idx)` to
-// `groovy_inspect_container`, which is a no-op on kinds other than
-// `ParenthesizedExpression` / `!`-prefixed `UnaryExpression`.
-fn groovy_inspect_child(node: &Node, idx: usize, conditions: &mut f64) {
-    if let Some(child) = node.child(idx) {
-        groovy_inspect_container(&child, conditions);
-    }
-}
-
-fn groovy_count_token_assignment(node: &Node, stats: &mut Stats) -> bool {
-    use Groovy::*;
-    match node.kind_id().into() {
-        STAREQ | SLASHEQ | PERCENTEQ | DASHEQ | PLUSEQ | LTLTEQ | GTGTEQ | AMPEQ | PIPEEQ
-        | CARETEQ | GTGTGTEQ | PLUSPLUS | DASHDASH => {
-            stats.assignments += 1.;
-        }
-        FieldDeclaration | LocalVariableDeclaration => {
-            stats.declaration.push(DeclKind::Var);
-        }
-        Final => {
-            if let Some(DeclKind::Var) = stats.declaration.last() {
-                stats.declaration.push(DeclKind::Const);
-            }
-        }
-        SEMI => {
-            if let Some(DeclKind::Const | DeclKind::Var) = stats.declaration.last() {
-                stats.declaration.clear();
-            }
-        }
-        EQ => {
-            if stats
-                .declaration
-                .last()
-                .is_none_or(|decl| matches!(decl, DeclKind::Var))
-            {
-                stats.assignments += 1.;
-            }
-        }
-        _ => return false,
-    }
-    true
-}
-
-fn groovy_count_token_branch(node: &Node, stats: &mut Stats) -> bool {
-    use Groovy::*;
-    if matches!(node.kind_id().into(), MethodInvocation | CommandChain | New) {
-        stats.branches += 1.;
-        return true;
-    }
-    false
-}
-
-// The `default` arm of a `switch` is excluded (issue #469): it is the
-// unconditional fallthrough, so cyclomatic counts only the `Case` arms
-// (Groovy shares Java's `impl_cyclomatic_java_like!`, which matches
-// `Case` and never `Default`).
-fn groovy_count_token_condition(node: &Node, stats: &mut Stats) -> bool {
-    use Groovy::*;
-    match node.kind_id().into() {
-        GTEQ | LTEQ | EQEQ | BANGEQ | Else | Case | QMARK | Try | Catch => {
-            stats.conditions += 1.;
-        }
-        // Excludes `<` / `>` used for generic types (e.g. `List<String>`).
-        GT | LT => {
-            if let Some(parent) = node.parent()
-                && !matches!(parent.kind_id().into(), TypeArguments)
-            {
-                stats.conditions += 1.;
-            }
-        }
-        _ => return false,
-    }
-    true
-}
-
-fn groovy_walk_for_conditions(node: &Node, stats: &mut Stats) {
-    use Groovy::*;
-    let conds = &mut stats.conditions;
-    match node.kind_id().into() {
-        AMPAMP | PIPEPIPE => {
-            if let Some(parent) = node.parent() {
-                groovy_count_unary_conditions(&parent, conds);
-            }
-        }
-        ArgumentList => groovy_count_unary_conditions(node, conds),
-        VariableDeclarator | AssignmentExpression => groovy_inspect_child(node, 2, conds),
-        // dekobon `if_statement` / `while_statement` shape:
-        // [keyword, `(`, condition, `)`, body, …]. Condition lives at
-        // child index 2 (not 1 as under tree-sitter-java, where parens
-        // wrap the condition in a `parenthesized_expression`).
-        IfStatement | WhileStatement => {
-            if let Some(condition) = node.child(2) {
-                groovy_count_condition(&condition, conds);
-            }
-        }
-        // dekobon shape: [`do`, body, `while`, `(`, condition, `)`].
-        // Condition is at child index 4.
-        DoWhileStatement => {
-            if let Some(condition) = node.child(4) {
-                groovy_count_condition(&condition, conds);
-            }
-        }
-        ReturnStatement => groovy_inspect_child(node, 1, conds),
-        TernaryExpression => groovy_walk_ternary(node, stats),
-        ForStatement => groovy_walk_for_statement(node, stats),
-        _ => {}
-    }
-}
-
-fn groovy_walk_ternary(node: &Node, stats: &mut Stats) {
-    let conds = &mut stats.conditions;
-    if let Some(condition) = node.child(0) {
-        groovy_count_condition(&condition, conds);
-    }
-    groovy_inspect_child(node, 2, conds);
-    groovy_inspect_child(node, 4, conds);
-}
-
-// Two shapes: a present condition lives at child(3); an empty condition
-// shows up as a bare `SEMI` token at child(3) with the next slot
-// (child(4)) holding either the update expression or `;` / `)` for
-// `for(;;)`-style loops, which still count as a single condition slot.
-fn groovy_walk_for_statement(node: &Node, stats: &mut Stats) {
-    use Groovy::*;
-    let Some(condition) = node.child(3) else {
-        return;
-    };
-    if !matches!(condition.kind_id().into(), SEMI) {
-        groovy_count_condition(&condition, &mut stats.conditions);
-        return;
-    }
-    let Some(cond) = node.child(4) else { return };
-    if matches!(cond.kind_id().into(), SEMI | RPAREN) {
-        stats.conditions += 1.;
-    } else {
-        groovy_count_condition(&cond, &mut stats.conditions);
-    }
-}
-
-impl Abc for GroovyCode {
-    // See `impl Abc for JavaCode` for the short-circuit-chain rationale
-    // and the cross-helper-exclusivity invariant.
-    fn compute<'a>(node: &Node<'a>, _code: &'a [u8], stats: &mut Stats) {
-        if groovy_count_token_assignment(node, stats) {
-            return;
-        }
-        if groovy_count_token_branch(node, stats) {
-            return;
-        }
-        if groovy_count_token_condition(node, stats) {
-            return;
-        }
-        groovy_walk_for_conditions(node, stats);
-    }
-}
-
-fn groovy_count_condition(condition: &Node, conditions: &mut f64) {
-    use Groovy::*;
-    // Terminal set mirrors the C# fix in #372 (lesson #19):
-    // `FieldAccess` (`obj.flag`), `CastExpression` (`v as Boolean` — the
-    // Groovy-idiomatic form), `ParenthesizedTypeCast` (`(boolean) v` —
-    // the Java-style form, which the dekobon Groovy grammar represents
-    // as its own kind rather than nesting `cast_expression` inside
-    // `parenthesized_expression`), and `InstanceofExpression`
-    // (`x instanceof Foo`) all evaluate to a boolean. The dekobon
-    // Groovy grammar has no `await` or `array_access` analogues, so
-    // those collapse out of the five-kind C# set.
-    match condition.kind_id().into() {
-        groovy_bool_terminal_kinds!() => {
-            *conditions += 1.;
-        }
-        ParenthesizedExpression | UnaryExpression => {
-            groovy_inspect_container(condition, conditions);
-        }
-        _ => {}
-    }
-}
-
-// ABC token-level helpers for C#. Mirror of Java's helper layout with
-// C#-specific deltas: every aliased kind id is matched via the
-// `csharp_*_kinds!()` macros (lesson #2); `ObjectCreationExpression`
-// joins `InvocationExpression*` as a branch; the `<` / `>` parent
-// guard widens to `TypeArgumentList | TypeParameterList |
-// FunctionPointerType`; `ConditionalExpression` replaces Java's
-// `TernaryExpression`; `for_statement` exposes its condition via the
-// named `condition` field rather than positional index.
-
-fn csharp_count_token_assignment(node: &Node, stats: &mut Stats) -> bool {
-    use Csharp::*;
-    match node.kind_id().into() {
-        STAREQ | SLASHEQ | PERCENTEQ | DASHEQ | PLUSEQ | LTLTEQ | GTGTEQ | GTGTGTEQ | AMPEQ
-        | PIPEEQ | CARETEQ | QMARKQMARKEQ | PLUSPLUS | DASHDASH => {
-            stats.assignments += 1.;
-        }
-        FieldDeclaration | LocalDeclarationStatement => {
-            stats.declaration.push(DeclKind::Var);
-        }
-        // C# `const` modifier marks a compile-time constant — exclude
-        // its initializer from the assignment count (matches Java's
-        // treatment of `final`).
-        Const => {
-            if let Some(DeclKind::Var) = stats.declaration.last() {
-                stats.declaration.push(DeclKind::Const);
-            }
-        }
-        SEMI => {
-            if let Some(DeclKind::Const | DeclKind::Var) = stats.declaration.last() {
-                stats.declaration.clear();
-            }
-        }
-        // Count `=` unless it's the initializer of a `const` declaration.
-        // `None` (outside any declaration) still counts.
-        EQ => {
-            if !matches!(stats.declaration.last(), Some(DeclKind::Const)) {
-                stats.assignments += 1.;
-            }
-        }
-        _ => return false,
-    }
-    true
-}
-
-fn csharp_count_token_branch(node: &Node, stats: &mut Stats) -> bool {
-    use Csharp::*;
-    if matches!(
-        node.kind_id().into(),
-        crate::Csharp::InvocationExpression
-            | crate::Csharp::InvocationExpression2
-            | crate::Csharp::InvocationExpression3
-            | ObjectCreationExpression
-    ) {
-        stats.branches += 1.;
-        return true;
-    }
-    false
-}
-
-fn csharp_count_token_condition(node: &Node, stats: &mut Stats) -> bool {
-    use Csharp::*;
-    match node.kind_id().into() {
-        // The statement `switch` counts its `Case` arms; the `default:`
-        // arm (token `Default`, shared by both classic `default:` and
-        // arrow `default ->` forms) is the unconditional fallthrough and
-        // is excluded, mirroring cyclomatic's `Case`-only count and the
-        // expression-arm discard rule below (issues #456, #469).
-        GTEQ | LTEQ | EQEQ | BANGEQ | Else | Case | QMARK | Try | Catch => {
-            stats.conditions += 1.;
-        }
-        // A `switch` *expression* arm (`x switch { 1 => …, _ => … }`) is a
-        // decision point. The statement `switch` counts via its `Case`
-        // tokens above; an expression arm carries none, so it scored zero
-        // conditions before #456 even though C# cyclomatic counts it. The
-        // bare-discard arm (`_ =>` / `var _ =>`, no `when` guard) is the
-        // `default:` analogue and is excluded — mirroring the cyclomatic
-        // gate (lesson 11).
-        SwitchExpressionArm
-            if !crate::metrics::cyclomatic::csharp_switch_expression_arm_is_bare_discard(node) =>
-        {
-            stats.conditions += 1.;
-        }
-        // Excludes `<` and `>` used as type-syntax delimiters: generic
-        // type arguments (`Dictionary<K, V>`), type parameter
-        // declarations (`class Foo<T> { }`), and the parameter-list
-        // delimiters of unsafe function-pointer types
-        // (`delegate*<int, int>`).
-        GT | LT => {
-            if let Some(parent) = node.parent()
-                && !matches!(
-                    parent.kind_id().into(),
-                    TypeArgumentList | TypeParameterList | FunctionPointerType
-                )
-            {
-                stats.conditions += 1.;
-            }
-        }
-        _ => return false,
-    }
-    true
-}
-
-fn csharp_walk_for_conditions(node: &Node, stats: &mut Stats) {
-    use Csharp::*;
-    let conds = &mut stats.conditions;
-    match node.kind_id().into() {
-        AMPAMP | PIPEPIPE => {
-            if let Some(parent) = node.parent() {
-                csharp_count_unary_conditions(&parent, conds);
-            }
-        }
-        ArgumentList => csharp_count_unary_conditions(node, conds),
-        // tree-sitter-c-sharp `if_statement` / `while_statement` shape:
-        // [`if`/`while`, `(`, condition, `)`, body, …]. The parens are
-        // anonymous string children, NOT a wrapping
-        // `parenthesized_expression` as in tree-sitter-java — so the
-        // condition lives at child(2). Targeting child(1) (the literal
-        // `(` token) was the #370 bug: every unary / bare-identifier
-        // condition silently scored 0. See issue #370.
-        IfStatement | WhileStatement => {
-            if let Some(condition) = node.child(2) {
-                csharp_count_condition(&condition, conds);
-            }
-        }
-        // tree-sitter-c-sharp `do_statement` shape:
-        // [`do`, body, `while`, `(`, condition, `)`, `;`]. The
-        // condition lives at child(4), not child(3) (which is the
-        // literal `(` token). Targeting child(3) was the second half
-        // of the #370 bug.
-        DoStatement => {
-            if let Some(condition) = node.child(4) {
-                csharp_count_condition(&condition, conds);
-            }
-        }
-        // `return value;` — child(1) is the value expression.
-        ReturnStatement => csharp_inspect_child(node, 1, conds),
-        // Child 2: declarator / assignment RHS, lambda body
-        // (`params => body`).
-        crate::Csharp::VariableDeclarator
-        | crate::Csharp::VariableDeclarator2
-        | AssignmentExpression
-        | LambdaExpression => csharp_inspect_child(node, 2, conds),
-        ConditionalExpression => csharp_walk_conditional(node, stats),
-        ForStatement => csharp_walk_for_statement(node, stats),
-        _ => {}
-    }
-}
-
-// `cond ? a : b` — children are [cond, ?, a, :, b]. The cond-classifier
-// match is shared with `csharp_walk_for_conditions`'s `if`/`while`/`do`
-// arms via `csharp_count_condition`; the two branch slots delegate to
-// `csharp_inspect_container` via `csharp_inspect_child` so a parenthesised
-// or `!`-prefixed branch contributes one condition just like a bare
-// invocation/identifier/boolean would.
-fn csharp_walk_conditional(node: &Node, stats: &mut Stats) {
-    let conds = &mut stats.conditions;
-    if let Some(condition) = node.child(0) {
-        csharp_count_condition(&condition, conds);
-    }
-    csharp_inspect_child(node, 2, conds);
-    csharp_inspect_child(node, 4, conds);
-}
-
-// Counts unary / single-token conditions inside `for` statements. The
-// C# grammar exposes the loop condition via the named `condition` field
-// on `for_statement`, so we look it up by name rather than positional
-// index. Comparison-operator conditions like `i < n` are still counted
-// by the standard `GT | LT | ...` arms — this only fires when the
-// condition is a bare identifier, invocation, boolean literal,
-// parenthesised expression, or `!`-prefixed unary expression.
-fn csharp_walk_for_statement(node: &Node, stats: &mut Stats) {
-    let Some(condition) = node.child_by_field_name("condition") else {
-        return;
-    };
-    let kind = condition.kind_id().into();
-    if matches!(kind, csharp_bool_terminal_kinds!()) {
-        stats.conditions += 1.;
-    } else if matches!(kind, csharp_paren_expr_kinds!())
-        || matches!(kind, csharp_prefix_unary_expr_kinds!())
-    {
-        csharp_inspect_container(&condition, &mut stats.conditions);
-    }
-}
-
-impl Abc for CsharpCode {
-    // See `impl Abc for JavaCode` for the short-circuit-chain rationale
-    // and the cross-helper-exclusivity invariant.
-    fn compute<'a>(node: &Node<'a>, _code: &'a [u8], stats: &mut Stats) {
-        if csharp_count_token_assignment(node, stats) {
-            return;
-        }
-        if csharp_count_token_branch(node, stats) {
-            return;
-        }
-        if csharp_count_token_condition(node, stats) {
-            return;
-        }
-        csharp_walk_for_conditions(node, stats);
-    }
-}
-
-// Elixir ABC unary-conditional walker (Fitzpatrick Rule 9; issue #557).
-// tree-sitter-elixir parses `a && b || c` as a left-nested chain of
-// `binary_operator` nodes (aliased `BinaryOperator`..`BinaryOperator3`
-// per lesson #2) carrying `&&` / `||` / `and` / `or` operator tokens.
-// Negation surfaces as `unary_operator` whose child(0) is the `!` token;
-// parenthesised operands parse as `block`. Both are unwrapped by
-// `elixir_inspect_container`.
-fn elixir_inspect_container(container_node: &Node, conditions: &mut f64) {
-    use Elixir as E;
-
-    let mut node = *container_node;
-    let mut node_kind = node.kind_id().into();
-    let Some(parent) = node.parent() else { return };
-    let mut has_boolean_content = matches!(
-        parent.kind_id().into(),
-        E::BinaryOperator | E::BinaryOperator2 | E::BinaryOperator3
-    );
-
-    loop {
-        let is_block = matches!(node_kind, E::Block);
-        let is_not = matches!(node_kind, E::UnaryOperator)
-            && node.child(0).is_some_and(|c| c.kind_id() == E::BANG as u16);
-
-        if !is_block && !is_not {
-            break;
-        }
-        if !has_boolean_content && is_not {
-            has_boolean_content = true;
-        }
-
-        // A `!` unary stores its operand at child index 1 (after the `!`
-        // token); a parenthesised `block` carries its inner expression as
-        // the first named child.
-        let next = if is_not {
-            node.child(1)
-        } else {
-            node.children().find(Node::is_named)
-        };
-        let Some(child) = next else { break };
-        node = child;
-        node_kind = node.kind_id().into();
-
-        if matches!(node_kind, elixir_bool_terminal_kinds!()) {
-            if has_boolean_content {
-                *conditions += 1.;
-            }
-            break;
-        }
-    }
-}
-
-// Counts each non-comparison operand of an Elixir `&&` / `||` chain once.
-// Comparison operands are nested `binary_operator` nodes (absent from
-// `elixir_bool_terminal_kinds!()`) and so contribute nothing.
-fn elixir_count_unary_conditions(list_node: &Node, conditions: &mut f64) {
-    use Elixir as E;
-
-    let list_kind = list_node.kind_id().into();
-    let mut cursor = list_node.cursor();
-
-    if cursor.goto_first_child() {
-        loop {
-            let node = cursor.node();
-            let node_kind = node.kind_id().into();
-
-            if matches!(node_kind, elixir_bool_terminal_kinds!())
-                && matches!(
-                    list_kind,
-                    E::BinaryOperator | E::BinaryOperator2 | E::BinaryOperator3
-                )
-            {
-                *conditions += 1.;
-            } else if node.is_named() {
-                elixir_inspect_container(&node, conditions);
-            }
-
-            if !cursor.goto_next_sibling() {
-                break;
-            }
-        }
-    }
-}
-
-impl Abc for ElixirCode {
-    // Elixir's pattern-match `=` is a `BinaryOperator` whose middle
-    // child is an `EQ` token. The same wrapper node also hosts `+=`-
-    // style augmented assignments, but Elixir is purely functional —
-    // augmented assignment does not exist in the grammar; `EQ` is the
-    // only assignment-shaped operator. `|>` (`PIPEGT`) is a
-    // BinaryOperator too but its operator token differs, so the EQ
-    // child check is what filters assignments from pipelines and from
-    // comparison operators that share the wrapper.
-    //
-    // Branches cover `|>` (the pipe operator dispatches one call per
-    // step) and every `Call` node (function / method / macro
-    // invocation). `RemoteCallWithParentheses` and `LocalCallWith*`
-    // variants are subordinate nodes to `Call`, so the single `Call`
-    // match captures every dispatch site.
-    //
-    // Conditions cover `when` (guard token `Elixir::When`), the six
-    // comparison operator tokens (`==`, `===`, `!=`, `!==`, `<`, `>`,
-    // `<=`, `>=`), and the keyword-shaped `Call`s that introduce a
-    // decision point (`if`, `unless`, `case`, `cond`, `with`).
-    // `for` / `while` are looping forms — not condition-shaped per
-    // the issue body's literal list — so we omit them.
-    //
-    // Limitations:
-    // - `case` is counted once on the container, not once per arm
-    //   (`stab_clause`). The issue body says "conditions = case,
-    //   cond, if, with, guard when" — i.e. one condition per
-    //   construct, not per arm. Matches the Rust impl's "MatchExpression
-    //   once" rule.
-    // - Higher-order calls like `Enum.reduce` are `RemoteCallWithParentheses`
-    //   nodes; they are still `Call` nodes and so contribute one branch
-    //   each, matching the issue's "branches = `|>`, function calls"
-    //   instruction.
-    fn compute<'a>(node: &Node<'a>, code: &'a [u8], stats: &mut Stats) {
-        use Elixir as E;
-
-        match node.kind_id().into() {
-            // A `BinaryOperator` whose operator token is `EQ` is a
-            // pattern-match assignment. The grammar shape is
-            // `(left, operator, right)`, so the operator token is
-            // always at child index 1 — looking it up directly is
-            // O(1) vs. an `any()` scan of all children. This arm
-            // fires on every Elixir binary op (comparisons, pipes,
-            // boolean ops, arithmetic) so the constant-time check
-            // matters.
-            E::BinaryOperator | E::BinaryOperator2 | E::BinaryOperator3
-                if node
-                    .child(1)
-                    .is_some_and(|c| c.kind_id() == E::EQ as u16) =>
-            {
-                stats.assignments += 1.;
-            }
-            // `|>` pipeline operator: every step in `foo |> bar |> baz`
-            // is one branch (the pipe dispatches one call per step).
-            E::PIPEGT => {
-                stats.branches += 1.;
-            }
-            // Every Call (function, method, macro, sigil-call) is one
-            // branch — `RemoteCallWith*`, `LocalCallWith*`,
-            // `AnonymousCall`, and `DoubleCall` are all subordinate
-            // node kinds underneath the top-level `Call` wrapper, so
-            // matching `Call` alone captures every dispatch site.
-            //
-            // Method-defining macros (`def`/`defp`/`defmacro`/`defmacrop`)
-            // and module/struct/protocol declarations (`defmodule`/
-            // `defstruct`/`defprotocol`/`defimpl`) are *not* runtime
-            // dispatch and must not inflate `branches` — they parse as
-            // `Call` nodes because Elixir's grammar uses the same
-            // shape for all keyword-introduced forms. Aliasing/import
-            // directives (`alias`, `import`, `require`, `use`) are
-            // similarly declarative and excluded.
-            //
-            // Cognitive's `elixir_call_keyword` lookup is reused to
-            // identify the target keyword. Note: Cognitive only acts
-            // on a subset of these keywords (the four method-definers
-            // for nesting reset, plus the 7 control-flow keywords for
-            // +nesting); Abc's broader filter additionally drops the
-            // module/struct/protocol declarators and aliasing
-            // directives that Cognitive ignores entirely. Filter sets
-            // are intentionally different — both impls use the same
-            // helper to look up the keyword, but apply different
-            // policies on top.
-            E::Call => {
-                let keyword = super::cognitive::elixir_call_keyword(node, code);
-                let is_definition_or_directive = matches!(
-                    keyword,
-                    Some(
-                        "def" | "defp" | "defmacro" | "defmacrop"
-                        | "defmodule" | "defstruct" | "defprotocol" | "defimpl"
-                        | "alias" | "import" | "require" | "use"
-                    )
-                );
-                if !is_definition_or_directive {
-                    stats.branches += 1.;
-                }
-                // Keyword-shaped control-flow Calls also contribute
-                // one condition.
-                if matches!(keyword, Some("if" | "unless" | "case" | "cond" | "with")) {
-                    stats.conditions += 1.;
-                }
-            }
-            // Comparison operator tokens. `Elixir::LT` / `Elixir::GT`
-            // are unambiguously comparison ops here — unlike Go's
-            // generic-instantiation `<` / `>`, Elixir has no type
-            // parameter brackets that share the token.
-            E::EQEQ | E::EQEQEQ | E::BANGEQ | E::BANGEQEQ
-            | E::LT | E::GT | E::LTEQ | E::GTEQ
-            // Guard `when` token: introduces the guard clause of a
-            // function head or `case` arm.
-            | E::When => {
-                stats.conditions += 1.;
-            }
-            // Fitzpatrick Rule 9 walker: each non-comparison operand of a
-            // `&&` / `||` / `and` / `or` chain is one condition (issue
-            // #557). The short-circuit operators are not counted directly
-            // (cross-language policy, #395); the keyword forms `and` / `or`
-            // get the same treatment as `&&` / `||`. Combined with the
-            // `if` Call already contributing one condition, `if a && b ||
-            // c` reports 4 — consistent with the cyclomatic count.
-            E::AMPAMP | E::PIPEPIPE | E::And | E::Or => {
-                if let Some(parent) = node.parent() {
-                    elixir_count_unary_conditions(&parent, &mut stats.conditions);
-                }
-            }
-            _ => {}
-        }
-    }
-}
-
-// Fitzpatrick's ABC rules adapted for Perl.
-//
-// - Assignments: every assignment operator token — plain `=` plus the
-//   compound forms `+=`, `-=`, `*=`, `/=`, `%=`, `**=`, `.=`, `x=`,
-//   `&=`, `|=`, `^=`, `<<=`, `>>=`, `&&=`, `||=`, `//=`, and the
-//   bitstring forms `&.=`, `|.=`, `^.=`. Each token fires exactly
-//   once per textual occurrence inside a `binary_expression`.
-// - Branches: every call expression dispatch — `call_expression_with_*`
-//   (bareword / spaced args / args-with-brackets / sub / variable /
-//   recursive) plus `method_invocation`. The grammar nests an inner
-//   `call_expression_with_bareword` (just the function name)
-//   underneath the wrapper kinds carrying argument lists, so we only
-//   count `CallExpressionWithBareword` when it stands on its own;
-//   when its parent is another call form, the outer wrapper has
-//   already contributed the branch.
-// - Conditions: numeric and string comparison operators (`==`, `!=`,
-//   `<`, `>`, `<=`, `>=`, `<=>`, `eq`, `ne`, `lt`, `gt`, `le`, `ge`,
-//   `cmp`, `=~`, `!~`), the ternary operator (`TernaryExpression`),
-//   and each `elsif` / `else` clause of an `if` / `unless`
-//   statement. Bare predicates that have no comparison (e.g.
-//   `if ($x)`) are not separately counted; we let the comparison
-//   tokens carry the metric, mirroring the Bash / Python token-
-//   level approach.
-//
-//   The short-circuit and low-precedence logical operators (`&&`,
-//   `||`, `//`, `and`, `or`, `xor`) are deliberately NOT counted.
-//   See the module-level `Stats` doc-comment for the cross-
-//   language policy (Fitzpatrick rules mapped from Figure 2 for C,
-//   the closest analogue since the paper does not define rules for
-//   Perl; issue #395, walker tracked in #403).
-// Perl ABC unary-conditional walker (Fitzpatrick Rule 9 mapped from
-// Figure 2 for C — the closest analogue, since the paper does not
-// define rules for Perl; issue #403). Logical-operator triggers cover
-// both the high-precedence punctuation (`&&`, `||`, `//`) and the
-// low-precedence keyword forms (`and`, `or`, `xor`). Terminal-bool
-// kinds: `Identifier`, `Boolean`, `True`, `False`, the call-expression
-// wrappers (every kind already counted as a branch), and the variable
-// wrappers (`ScalarVariable`, `ArrayVariable`, `HashVariable` plus the
-// access shapes).
-fn perl_inspect_container(container_node: &Node, conditions: &mut f64) {
-    use Perl as P;
-
-    let mut node = *container_node;
-    let mut node_kind = node.kind_id().into();
-    let Some(parent) = node.parent() else { return };
-    let parent_kind = parent.kind_id().into();
-    let mut has_boolean_content = matches!(
-        parent_kind,
-        P::BinaryExpression
-            | P::IfStatement
-            | P::UnlessStatement
-            | P::WhileStatement
-            | P::UntilStatement
-    ) || (matches!(parent_kind, P::TernaryExpression)
-        && node
-            .previous_sibling()
-            .is_none_or(|prev| !matches!(prev.kind_id().into(), P::QMARK | P::COLON)));
-
-    loop {
-        // `Array` is tree-sitter-perl's name for the `(...)` shape
-        // used BOTH as the if/while/unless/until condition wrapper
-        // AND as list literals `(1, 2, 3)` (and `(x, y)` operand
-        // groupings). In Perl's scalar context — which every walker
-        // call site here operates in — a list expression evaluates
-        // to its LAST element, so descending via the last named
-        // child gives the semantically correct operand for both
-        // shapes: `($a)` → `$a`, `($x, $y)` → `$y`, `if ($a)` →
-        // `$a`. `ParenthesizedArgument` (the other paren-wrap kind)
-        // has only one inner expression, so child(1) and last-named
-        // are equivalent.
-        let is_parens = matches!(node_kind, P::ParenthesizedArgument | P::Array);
-        let is_not = matches!(node_kind, P::UnaryExpression)
-            && node.child(0).is_some_and(|c| c.kind_id() == P::BANG as u16);
-
-        if !is_parens && !is_not {
-            break;
-        }
-        if !has_boolean_content && is_not {
-            has_boolean_content = true;
-        }
-
-        // Descend through the wrapper to the value. Array uses
-        // last-named-child (Perl scalar-context value); other
-        // wrappers store their inner expression at child(1).
-        let next = if matches!(node_kind, P::Array) {
-            perl_last_named_child(&node)
-        } else {
-            node.child(1)
-        };
-        let Some(child) = next else { break };
-        node = child;
-        node_kind = node.kind_id().into();
-
-        if matches!(node_kind, perl_bool_terminal_kinds!()) {
-            if has_boolean_content {
-                *conditions += 1.;
-            }
-            break;
-        }
-    }
-}
-
-// Phase-2B (issue #403): pass `node.child(idx)` through
-// `perl_inspect_container`. Perl wraps `if (cond)` / `while (cond)` /
-// `unless (cond)` / `until (cond)` conditions in a
-// `parenthesized_argument`, so the paren unwrap handles the
-// boolean-literal case.
-fn perl_inspect_child(node: &Node, idx: usize, conditions: &mut f64) {
-    if let Some(child) = node.child(idx) {
-        perl_inspect_container(&child, conditions);
-    }
-}
-
-// Phase-2B helper (issue #403): Perl's `Array` node serves double
-// duty as the `(...)` wrapper around `if` / `while` / `unless` /
-// `until` conditions AND as the call-argument-list wrapper. The
-// dispatcher routes call-argument Arrays through
-// `perl_count_unary_conditions`; condition-slot Arrays are
-// already unwrapped by `perl_inspect_container`. This predicate
-// disambiguates by checking the parent kind.
-// Returns the last named child of a node, or None if there are no
-// named children. Used by `perl_inspect_container` to descend through
-// the `Array` `(...)` wrapper: for a single-element grouping
-// `($a)` the last named child is `$a`; for a multi-element list
-// literal `($x, $y)` the last named child is `$y` (the value the
-// expression evaluates to in Perl's scalar context, which is the
-// only context the walker operates in).
-fn perl_last_named_child<'a>(node: &Node<'a>) -> Option<Node<'a>> {
-    let mut cursor = node.cursor();
-    let mut last_named = None;
-    if cursor.goto_first_child() {
-        loop {
-            let child = cursor.node();
-            if child.is_named() {
-                last_named = Some(child);
-            }
-            if !cursor.goto_next_sibling() {
-                break;
-            }
-        }
-    }
-    last_named
-}
-
-fn perl_is_call_argument_parent(parent: Node) -> bool {
-    use Perl as P;
-    matches!(
-        parent.kind_id().into(),
-        P::CallExpressionWithArgsWithBrackets
-            | P::CallExpressionWithSpacedArgs
-            | P::CallExpressionWithSub
-            | P::CallExpressionWithVariable
-            | P::CallExpressionRecursive
-            | P::MethodInvocation
-    )
-}
-
-fn perl_count_unary_conditions(list_node: &Node, conditions: &mut f64) {
-    use Perl as P;
-
-    let list_kind = list_node.kind_id().into();
-    let mut cursor = list_node.cursor();
-
-    if cursor.goto_first_child() {
-        loop {
-            let node = cursor.node();
-            let node_kind = node.kind_id().into();
-
-            if matches!(node_kind, perl_bool_terminal_kinds!())
-                && matches!(list_kind, P::BinaryExpression)
-            {
-                *conditions += 1.;
-            } else if node.is_named() {
-                perl_inspect_container(&node, conditions);
-            }
-
-            if !cursor.goto_next_sibling() {
-                break;
-            }
-        }
-    }
-}
-
-impl Abc for PerlCode {
-    fn compute<'a>(node: &Node<'a>, _code: &'a [u8], stats: &mut Stats) {
-        use Perl as P;
-
-        match node.kind_id().into() {
-            // Plain `=` and every compound assignment operator. The
-            // grammar tokenises each operator separately, so one
-            // textual `+=` produces exactly one token and there is no
-            // double-counting via a wrapper.
-            P::EQ
-            | P::PLUSEQ
-            | P::DASHEQ
-            | P::STAREQ
-            | P::SLASHEQ
-            | P::PERCENTEQ
-            | P::STARSTAREQ
-            | P::DOTEQ
-            | P::XEQ
-            | P::AMPEQ
-            | P::PIPEEQ
-            | P::CARETEQ
-            | P::LTLTEQ
-            | P::GTGTEQ
-            | P::AMPAMPEQ
-            | P::PIPEPIPEEQ
-            | P::SLASHSLASHEQ
-            | P::AMPDOTEQ
-            | P::PIPEDOTEQ
-            | P::CARETDOTEQ => {
-                stats.assignments += 1.;
-            }
-            // Argument-bearing call wrappers always count.
-            P::CallExpressionWithSpacedArgs
-            | P::CallExpressionWithSub
-            | P::CallExpressionWithArgsWithBrackets
-            | P::CallExpressionWithVariable
-            | P::CallExpressionRecursive
-            | P::MethodInvocation => {
-                stats.branches += 1.;
-            }
-            // Bareword-only call (`shift`, `time`, …) — count only
-            // when this node is the outermost dispatch site. When the
-            // bareword sits inside one of the wrappers above, the
-            // outer node has already been counted and this child
-            // would double the branch tally.
-            P::CallExpressionWithBareword
-                if !node.parent().is_some_and(|p| {
-                    matches!(
-                        p.kind_id().into(),
-                        P::CallExpressionWithSpacedArgs
-                            | P::CallExpressionWithSub
-                            | P::CallExpressionWithArgsWithBrackets
-                            | P::CallExpressionWithVariable
-                            | P::CallExpressionRecursive
-                    )
-                }) =>
-            {
-                stats.branches += 1.;
-            }
-            // Numeric, string, and pattern-match comparison operators
-            // plus the spaceship / `cmp` three-way comparisons.
-            P::EQEQ | P::BANGEQ | P::LT | P::GT | P::LTEQ | P::GTEQ | P::LTEQGT
-            | P::Eq | P::Ne | P::Lt | P::Gt | P::Le | P::Ge | P::Cmp
-            | P::EQTILDE | P::BANGTILDE
-            // Ternary `a ? b : c` and each `elsif` / `else` clause of
-            // an `if` / `unless` chain.
-            | P::TernaryExpression
-            | P::ElsifClause
-            | P::ElseClause => {
-                stats.conditions += 1.;
-            }
-            // Fitzpatrick Rule 9 walker: each operand of a Perl
-            // short-circuit / low-precedence logical chain is one
-            // condition (issue #403). Covers `&&`, `||`, `//`,
-            // `and`, `or`, `xor`.
-            P::AMPAMP | P::PIPEPIPE | P::SLASHSLASH | P::And | P::Or | P::Xor => {
-                if let Some(parent) = node.parent() {
-                    perl_count_unary_conditions(&parent, &mut stats.conditions);
-                }
-            }
-            // Phase-2B (issue #403): condition slots. Perl wraps
-            // `if (cond)` / `while (cond)` / `unless (cond)` /
-            // `until (cond)` in the `Array` `(...)` shape (the
-            // grammar's name for parenthesized expressions in
-            // statement-modifier slots) — the paren unwrap handles
-            // boolean-literal cases. Condition sits at child(1)
-            // (child(0) is the `if` / `while` keyword).
-            // `return value`'s value also sits at child(1); merged
-            // into the same arm body to satisfy `match_same_arms`.
-            P::IfStatement
-            | P::UnlessStatement
-            | P::WhileStatement
-            | P::UntilStatement
-            | P::ReturnExpression => {
-                perl_inspect_child(node, 1, &mut stats.conditions);
-            }
-            // `call(!$a, !$b)` — argument list walker. Perl wraps
-            // call-argument lists in an `Array` node (same kind name
-            // as the `(...)` wrapper around `if` / `while`
-            // conditions). To avoid re-handling condition slots that
-            // were already walked through inspect_container, only
-            // dispatch when the parent is a call-expression form.
-            P::Array if node.parent().is_some_and(perl_is_call_argument_parent) => {
-                perl_count_unary_conditions(node, &mut stats.conditions);
-            }
-            _ => {}
-        }
-    }
-}
-
-// Fitzpatrick's ABC rules adapted for Lua.
-//
-// - Assignments: every `assignment_statement` node. Lua has no
-//   compound assignment operators (`+=` and friends do not exist in
-//   the grammar), so the wrapper kind is the sole assignment node
-//   and there is no per-operator alternative to track. `local x = 1`
-//   wraps an `assignment_statement` under a `variable_declaration`,
-//   so initialisers count the same as later mutations.
-// - Branches: every `function_call`. The Lua grammar collapses
-//   `obj.method(args)`, `obj:method(args)`, and `f(args)` into the
-//   same `function_call` node, so one arm covers all dispatch forms.
-// - Conditions: comparison operators (`==`, `~=`, `<`, `>`, `<=`,
-//   `>=`) and each elseif / else arm of an `if`. Lua has no ternary
-//   operator (`cond and a or b` is the idiom). The short-circuit
-//   operators `and` / `or` are deliberately NOT counted; see the
-//   module-level `Stats` doc-comment for the cross-language policy
-//   (issue #395, walker tracked in #403).
-// Lua ABC unary-conditional walker (Fitzpatrick Rule 9; issue #403).
-// Lua's logical operators are keyword tokens (`and` / `or`) inside a
-// `binary_expression`; `not x` is a `unary_expression` whose first
-// child is the `not` keyword. Terminal-bool kinds include identifiers,
-// the three keyword literals (`true`, `false`, `nil`), numbers, and
-// every call / indexing form.
-fn lua_inspect_container(container_node: &Node, conditions: &mut f64) {
-    let mut node = *container_node;
-    let mut node_kind = node.kind_id().into();
-    let Some(parent) = node.parent() else { return };
-    let mut has_boolean_content = matches!(
-        parent.kind_id().into(),
-        Lua::BinaryExpression | Lua::IfStatement | Lua::WhileStatement | Lua::RepeatStatement
-    );
-
-    loop {
-        let is_parens = matches!(node_kind, Lua::ParenthesizedExpression);
-        let is_not = matches!(node_kind, Lua::UnaryExpression)
-            && node
-                .child(0)
-                .is_some_and(|c| c.kind_id() == Lua::Not as u16);
-
-        if !is_parens && !is_not {
-            break;
-        }
-        // A `not` wrapper proves the operand is boolean even when
-        // the parent context didn't — matches the JS/Java/etc.
-        // pattern. Without this, `m(not a)` and other call-argument
-        // contexts would never propagate has_boolean_content.
-        has_boolean_content |= is_not;
-
-        let Some(child) = node.child(1) else { break };
-        node = child;
-        node_kind = node.kind_id().into();
-
-        if matches!(node_kind, lua_bool_terminal_kinds!()) {
-            if has_boolean_content {
-                *conditions += 1.;
-            }
-            break;
-        }
-    }
-}
-
-// Phase-2B (issue #403): Lua `if` / `while` / `repeat` condition
-// slots. Lua has no paren wrap, so the condition has to be classified
-// directly: terminal-bool kinds (Identifier, True, False, Nil,
-// FunctionCall, etc.) count at the top level; `(...)` / `not ...`
-// route through `lua_inspect_container`.
-fn lua_count_condition(condition: &Node, conditions: &mut f64) {
-    let kind = condition.kind_id().into();
-    if matches!(kind, lua_bool_terminal_kinds!()) {
-        *conditions += 1.;
-    } else if matches!(kind, Lua::ParenthesizedExpression | Lua::UnaryExpression) {
-        lua_inspect_container(condition, conditions);
-    }
-}
-
-fn lua_count_unary_conditions(list_node: &Node, conditions: &mut f64) {
-    let list_kind = list_node.kind_id().into();
-    let mut cursor = list_node.cursor();
-
-    if cursor.goto_first_child() {
-        loop {
-            let node = cursor.node();
-            let node_kind = node.kind_id().into();
-
-            if matches!(node_kind, lua_bool_terminal_kinds!())
-                && matches!(list_kind, Lua::BinaryExpression)
-            {
-                *conditions += 1.;
-            } else if node.is_named() {
-                lua_inspect_container(&node, conditions);
-            }
-
-            if !cursor.goto_next_sibling() {
-                break;
-            }
-        }
-    }
-}
-
-impl Abc for LuaCode {
-    fn compute<'a>(node: &Node<'a>, _code: &'a [u8], stats: &mut Stats) {
-        match node.kind_id().into() {
-            Lua::AssignmentStatement | Lua::AssignmentStatement2 => {
-                stats.assignments += 1.;
-            }
-            Lua::FunctionCall => {
-                stats.branches += 1.;
-            }
-            Lua::EQEQ
-            | Lua::TILDEEQ
-            | Lua::LT
-            | Lua::GT
-            | Lua::LTEQ
-            | Lua::GTEQ
-            | Lua::ElseifStatement
-            | Lua::ElseStatement => {
-                stats.conditions += 1.;
-            }
-            // Fitzpatrick Rule 9 walker: each operand of an `and` /
-            // `or` chain is one condition (issue #403).
-            Lua::And | Lua::Or => {
-                if let Some(parent) = node.parent() {
-                    lua_count_unary_conditions(&parent, &mut stats.conditions);
-                }
-            }
-            // Phase-2B (issue #403): condition slots. Lua has no paren
-            // wrap around `if` / `while` / `repeat …` conditions, so
-            // `lua_count_condition` classifies the slot directly. Use
-            // `child_by_field_name("condition")` so the lookup is
-            // grammar-version-robust — tree-sitter-lua exposes the
-            // `condition` field on if/while/repeat statements. Pinning
-            // by name handles the rare empty-body `repeat until cond`
-            // shape where the BLANK alternative for the body would
-            // shift positional child indices.
-            Lua::IfStatement | Lua::WhileStatement | Lua::RepeatStatement => {
-                if let Some(cond) = node.child_by_field_name("condition") {
-                    lua_count_condition(&cond, &mut stats.conditions);
-                }
-            }
-            // `return value` — Lua wraps return values in
-            // `expression_list` at child(1). Route each named child
-            // through `inspect_container` (no top-level terminal
-            // count) so `return not x` counts the unary unwrap once
-            // while `return x` (bare) reports zero. Bare `return`
-            // (no values) has no child(1).
-            Lua::ReturnStatement => {
-                if let Some(expr_list) = node.child(1) {
-                    for_each_named_child(&expr_list, &mut stats.conditions, lua_inspect_container);
-                }
-            }
-            // `f(not a, not b)` — argument-list walker.
-            Lua::Arguments => {
-                lua_count_unary_conditions(node, &mut stats.conditions);
-            }
-            _ => {}
-        }
-    }
-}
-
-// Names of Tcl commands that mutate a variable. Each invocation of
-// one of these commands counts as an assignment, not a branch — the
-// command is acting as an assignment operator, not as a generic
-// dispatch. The list is intentionally narrow: only commands that
-// every Tcl programmer recognises as primary mutators. Less-common
-// mutators (`dict set`, `array set`, `lset`, `regsub … name`) are
-// left as branches; treating them as assignments would require
-// inspecting the command's second word, and the additional
-// fidelity is not worth the complexity for the ABC magnitude.
-const TCL_ASSIGNMENT_COMMANDS: &[&[u8]] = &[b"incr", b"append", b"lappend"];
-
-// Fitzpatrick's ABC rules adapted for Tcl.
-//
-// - Assignments: every `set` production (`set name value`) plus
-//   every `command` whose first word is one of the recognised
-//   mutator commands in `TCL_ASSIGNMENT_COMMANDS`. Tcl has no
-//   assignment operators — variable mutation is always a command
-//   invocation, so we filter on the command name. The `set` form
-//   has its own grammar production (`Tcl::Set`) and counts directly
-//   without any source-text inspection.
-// - Branches: every other `command` node. Like Bash, `return` and
-//   `error` builtins parse as plain `command` nodes and count here
-//   too — Tcl treats every dispatch the same regardless of whether
-//   the command is a procedure call, a control-flow primitive, or a
-//   builtin. The grammar productions for `if`, `while`, `foreach`,
-//   etc. live separately from `command` and do not double-count.
-// - Conditions: numeric (`==`, `!=`, `<`, `>`, `<=`, `>=`) and
-//   string (`eq`, `ne`, `in`, `ni`) comparison tokens, the ternary
-//   expression production, and each `elseif` / `else` clause of an
-//   `if`. The short-circuit operators `&&` / `||` are deliberately
-//   NOT counted; see the module-level `Stats` doc-comment for the
-//   cross-language policy (issue #395, walker tracked in #403).
-// Tcl ABC unary-conditional walker (Fitzpatrick Rule 9; issue #403).
-// Tcl expression syntax appears inside `if {…}` / `while {…}` braces
-// and parses as `binop_expr` whose operator tokens include `AMPAMP`
-// and `PIPEPIPE`. Terminal-bool kinds are the bare-word literals
-// `simple_word`, the braced / quoted variants, variable substitutions
-// (`$x`), command substitutions (`[cmd]`), the boolean keyword, and
-// the numeric literal.
-fn tcl_inspect_container(container_node: &Node, conditions: &mut f64) {
-    let mut node = *container_node;
-    let mut node_kind = node.kind_id().into();
-    let Some(parent) = node.parent() else { return };
-    let has_boolean_content = matches!(parent.kind_id().into(), Tcl::BinopExpr);
-
-    loop {
-        let is_not = matches!(node_kind, Tcl::UnaryExpr)
-            && node
-                .child(0)
-                .is_some_and(|c| c.kind_id() == Tcl::BANG as u16);
-
-        if !is_not {
-            break;
-        }
-
-        let Some(child) = node.child(1) else { break };
-        node = child;
-        node_kind = node.kind_id().into();
-
-        if matches!(node_kind, tcl_bool_terminal_kinds!()) {
-            if has_boolean_content {
-                *conditions += 1.;
-            }
-            break;
-        }
-    }
-}
-
-fn tcl_count_unary_conditions(list_node: &Node, conditions: &mut f64) {
-    let list_kind = list_node.kind_id().into();
-    let mut cursor = list_node.cursor();
-
-    if cursor.goto_first_child() {
-        loop {
-            let node = cursor.node();
-            let node_kind = node.kind_id().into();
-
-            if matches!(node_kind, tcl_bool_terminal_kinds!())
-                && matches!(list_kind, Tcl::BinopExpr)
-            {
-                *conditions += 1.;
-            } else if node.is_named() {
-                tcl_inspect_container(&node, conditions);
-            }
-
-            if !cursor.goto_next_sibling() {
-                break;
-            }
-        }
-    }
-}
-
-impl Abc for TclCode {
-    fn compute<'a>(node: &Node<'a>, code: &'a [u8], stats: &mut Stats) {
-        match node.kind_id().into() {
-            // The `set` production wraps `set name value` as a
-            // first-class node distinct from generic commands.
-            Tcl::Set => {
-                stats.assignments += 1.;
-            }
-            // Generic command: branch by default, assignment when
-            // the first word names a known mutator. The first word
-            // can be either a `simple_word` or a wrapped form; both
-            // surface their literal text via `utf8_text`.
-            Tcl::Command => {
-                if tcl_command_is_assignment(node, code) {
-                    stats.assignments += 1.;
-                } else {
-                    stats.branches += 1.;
-                }
-            }
-            Tcl::EQEQ
-            | Tcl::BANGEQ
-            | Tcl::LT
-            | Tcl::GT
-            | Tcl::LTEQ
-            | Tcl::GTEQ
-            | Tcl::Eq
-            | Tcl::Ne
-            | Tcl::In
-            | Tcl::Ni
-            | Tcl::TernaryExpr
-            | Tcl::Elseif
-            | Tcl::Else => {
-                stats.conditions += 1.;
-            }
-            // Fitzpatrick Rule 9 walker: each operand of a `&&` / `||`
-            // chain inside an `expr` slot is one condition (issue #403).
-            Tcl::AMPAMP | Tcl::PIPEPIPE => {
-                if let Some(parent) = node.parent() {
-                    tcl_count_unary_conditions(&parent, &mut stats.conditions);
-                }
-            }
-            _ => {}
-        }
-    }
-}
-
-impl Abc for IrulesCode {
-    fn compute<'a>(node: &Node<'a>, code: &'a [u8], stats: &mut Stats) {
-        match node.kind_id().into() {
-            // The `set name value` production is a first-class node.
-            Irules::Set => {
-                stats.assignments += 1.;
-            }
-            // Generic command: assignment when the first word names a known
-            // mutator (`incr`/`append`/`lappend`), otherwise a branch — every
-            // dispatch counts, including `return`. The `if`/`while`/`switch`/…
-            // productions are separate kinds and do not reach this arm.
-            Irules::Command => {
-                if irules_command_is_assignment(node, code) {
-                    stats.assignments += 1.;
-                } else {
-                    stats.branches += 1.;
-                }
-            }
-            // Numeric and string comparison tokens, the ternary expression,
-            // and each `elseif` / `else` clause. iRules adds the word-form
-            // string comparators (`starts_with`, `contains`, `matches`, …)
-            // that Tcl lacks.
-            Irules::EQEQ
-            | Irules::BANGEQ
-            | Irules::LT
-            | Irules::GT
-            | Irules::LTEQ
-            | Irules::GTEQ
-            | Irules::Eq
-            | Irules::Ne
-            | Irules::StartsWith
-            | Irules::EndsWith
-            | Irules::Contains
-            | Irules::Equals
-            | Irules::Matches
-            | Irules::MatchesRegex
-            | Irules::MatchesGlob
-            | Irules::In
-            | Irules::Ni
-            | Irules::TernaryExpr
-            | Irules::Elseif
-            | Irules::Else => {
-                stats.conditions += 1.;
-            }
-            // Fitzpatrick Rule 9: the short-circuit operators are not counted
-            // directly (cross-language policy, #395); instead each operand of
-            // a `&&`/`||`/`and`/`or` chain is one condition (#403). iRules'
-            // keyword forms (`and`/`or`) get the same treatment as `&&`/`||`.
-            Irules::AMPAMP | Irules::PIPEPIPE | Irules::And | Irules::Or => {
-                if let Some(parent) = node.parent() {
-                    irules_count_unary_conditions(&parent, &mut stats.conditions);
-                }
-            }
-            _ => {}
-        }
-    }
-}
-
-// Returns true when the `command` node's first word is one of the
-// recognised Tcl assignment commands. The first word is the leftmost
-// non-comment child; we slice the source bytes directly using the
-// child node's byte range, which is robust to `simple_word` wrappers
-// and avoids depending on a particular grammar shape.
-fn tcl_command_is_assignment(node: &Node, code: &[u8]) -> bool {
-    let Some(first) = node.child(0) else {
-        return false;
-    };
-    let start = first.start_byte();
-    let end = first.end_byte();
-    if end > code.len() || start >= end {
-        return false;
-    }
-    let word = &code[start..end];
-    TCL_ASSIGNMENT_COMMANDS.contains(&word)
-}
-
-// iRules mutator commands (same Tcl builtins; the dedicated `set`
-// production is handled separately in the impl, like Tcl).
-const IRULES_ASSIGNMENT_COMMANDS: &[&[u8]] = &[b"incr", b"append", b"lappend"];
-
-// iRules counterpart of `tcl_command_is_assignment`.
-fn irules_command_is_assignment(node: &Node, code: &[u8]) -> bool {
-    let Some(first) = node.child(0) else {
-        return false;
-    };
-    let start = first.start_byte();
-    let end = first.end_byte();
-    if end > code.len() || start >= end {
-        return false;
-    }
-    let word = &code[start..end];
-    IRULES_ASSIGNMENT_COMMANDS.contains(&word)
-}
-
-// iRules counterpart of `tcl_inspect_container` (Fitzpatrick Rule 9): a
-// negated bare operand (`!$flag`) inside a boolean chain is one condition.
-fn irules_inspect_container(container_node: &Node, conditions: &mut f64) {
-    let mut node = *container_node;
-    let mut node_kind = node.kind_id().into();
-    let Some(parent) = node.parent() else { return };
-    let has_boolean_content = matches!(parent.kind_id().into(), Irules::BinopExpr);
-
-    loop {
-        let is_not = matches!(node_kind, Irules::UnaryExpr)
-            && node
-                .child(0)
-                .is_some_and(|c| c.kind_id() == Irules::BANG as u16);
-
-        if !is_not {
-            break;
-        }
-
-        let Some(child) = node.child(1) else { break };
-        node = child;
-        node_kind = node.kind_id().into();
-
-        if matches!(node_kind, irules_bool_terminal_kinds!()) {
-            if has_boolean_content {
-                *conditions += 1.;
-            }
-            break;
-        }
-    }
-}
-
-// iRules counterpart of `tcl_count_unary_conditions`.
-fn irules_count_unary_conditions(list_node: &Node, conditions: &mut f64) {
-    let list_kind = list_node.kind_id().into();
-    let mut cursor = list_node.cursor();
-
-    if cursor.goto_first_child() {
-        loop {
-            let node = cursor.node();
-            let node_kind = node.kind_id().into();
-
-            if matches!(node_kind, irules_bool_terminal_kinds!())
-                && matches!(list_kind, Irules::BinopExpr)
-            {
-                *conditions += 1.;
-            } else if node.is_named() {
-                irules_inspect_container(&node, conditions);
-            }
-
-            if !cursor.goto_next_sibling() {
-                break;
-            }
-        }
-    }
-}
-
-// C# mirror of `java_inspect_child` / `groovy_inspect_child`: passes
-// `node.child(idx)` to `csharp_inspect_container`, which is a no-op on
-// kinds other than `csharp_paren_expr_kinds!()` / `!`-prefixed
-// `csharp_prefix_unary_expr_kinds!()`.
-fn csharp_inspect_child(node: &Node, idx: usize, conditions: &mut f64) {
-    if let Some(child) = node.child(idx) {
-        csharp_inspect_container(&child, conditions);
-    }
-}
-
-fn csharp_count_condition(condition: &Node, conditions: &mut f64) {
-    let kind = condition.kind_id().into();
-    if matches!(kind, csharp_bool_terminal_kinds!()) {
-        *conditions += 1.;
-    } else if matches!(kind, csharp_paren_expr_kinds!())
-        || matches!(kind, csharp_prefix_unary_expr_kinds!())
-    {
-        csharp_inspect_container(condition, conditions);
-    }
-}
 
 #[cfg(test)]
 #[allow(
